% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_long_pipeline.R
\name{addRowRanges}
\alias{addRowRanges}
\title{Add rowRanges by rownames to \code{SummarizedExperiment} object
Assumes rownames are transcript_ids
Assumes transcript_id is present in the annotation file}
\usage{
addRowRanges(sce, annotation, outdir)
}
\value{
a \code{SummarizedExperiment} object with rowRanges added
}
\description{
Add rowRanges by rownames to \code{SummarizedExperiment} object
Assumes rownames are transcript_ids
Assumes transcript_id is present in the annotation file
}
\keyword{internal}
