% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_reads_to_genome.R
\name{align_reads_to_genome}
\alias{align_reads_to_genome}
\title{Title}
\usage{
align_reads_to_genome(
  gene_anno,
  input_fq,
  genome_fa,
  genome_bam,
  minimap2_dir,
  out_dir,
  use_junctions = TRUE,
  no_flank = TRUE
)
}
\arguments{
\item{genome_bam}{Output Bam file, MORE}

\item{name}{desc}
}
\description{
DESC
}
