% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gff.R
\name{annotate_filter_gff}
\alias{annotate_filter_gff}
\title{Title}
\usage{
annotate_filter_gff(
  isoform_gff,
  ref_gff,
  isoform_out,
  anno_out,
  tr_cnt,
  min_sup_reads
)
}
\arguments{
\item{isoform_gff}{NEEDED}

\item{ref_gff}{NEEDED}

\item{min_sup_reads}{NEEDED}
}
\description{
Combine FLAMES ouput with reference and filter out transcript by
ealignment result.
}
