% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{annotation_to_fasta}
\alias{annotation_to_fasta}
\title{GTF/GFF to FASTA conversion}
\usage{
annotation_to_fasta(isoform_annotation, genome_fa, outdir)
}
\arguments{
\item{isoform_annotation}{Path to the annotation file (GTF/GFF3)}

\item{genome_fa}{The file path to genome fasta file.}

\item{outdir}{The path to directory to store the transcriptome as \code{transcript_assembly.fa}.}
}
\value{
Path to the outputted transcriptome assembly
}
\description{
convert the transcript annotation to transcriptome assembly as FASTA file.
}
\examples{
fasta <- annotation_to_fasta(system.file("extdata/rps24.gtf.gz", package = "FLAMES"), system.file("extdata/rps24.fa.gz", package = "FLAMES"), tempdir())
cat(readChar(fasta, nchars = 1e3))

}
