% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_barcode.R
\name{barcode_info_plots}
\alias{barcode_info_plots}
\title{Barcode demultiplexing QC plots}
\usage{
barcode_info_plots(sce)
}
\arguments{
\item{sce}{The \code{SingleCellExperiment} object from FLAMES pipeline, or the returned list from \code{find_barcode}}
}
\value{
a list of QC plots for the barcode demultiplexing step (\code{find_barcode})
}
\description{
Plot the barcode demultiplexing statistics
}
\examples{
outdir <- tempfile()
dir.create(outdir)
bc_allow <- file.path(outdir, 'bc_allow.tsv')
R.utils::gunzip(filename = system.file('extdata/bc_allow.tsv.gz', package = 'FLAMES'), destname = bc_allow, remove = FALSE)
barcode_info <- find_barcode(
   fastq_dir = system.file('extdata/fastq', package = 'FLAMES'),
   stats_file = file.path(outdir, 'bc_stat'),
   out_fastq = file.path(outdir, 'demultiplexed.fq.gz'),
   ref_csv = bc_allow,
   MAX_DIST = 2,
   UMI_LEN = 10
)
barcode_info_plots(barcode_info)
}
