% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bernoulli_like}
\alias{bernoulli_like}
\title{Bernoulli likelihood}
\usage{
bernoulli_like(y, X, p_beta, p_c, p_df, inv_link_f, log = TRUE)
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{p_beta}{Regression coefficients}

\item{p_c}{c parameter}

\item{p_df}{Degrees of freedom}

\item{inv_link_f}{Inverse link function}

\item{log}{For log-scale values}
}
\value{
Likelihood value
}
\description{
Bernoulli likelihood
}
