% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_posterior}
\alias{beta_posterior}
\title{Beta posterior distribution}
\usage{
beta_posterior(y, X, p_beta, p_beta_element, element, p_c, p_df,
  inv_link_f, sigma_beta = 100, log = TRUE, method = "ARMS")
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{p_beta}{Regression coefficients}

\item{p_beta_element}{Specific coefficient element}

\item{element}{Element you are sampling}

\item{p_c}{c parameter}

\item{p_df}{Degrees of freedom}

\item{inv_link_f}{Inverse link function}

\item{sigma_beta}{Variance of beta[element] prior}

\item{log}{For log-scale values}

\item{method}{"ARMS" or "metropolis"}
}
\value{
Likelihood value
}
\description{
Beta posterior distribution
}
