% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{blocks_to_junctions}
\alias{blocks_to_junctions}
\title{Blocks To Junctions}
\usage{
blocks_to_junctions(block)
}
\arguments{
\item{block}{NEEDED}
}
\value{
The converted junctions named list
}
\description{
Converts \code{block} to a named list containing the keys "left", "right"
and "junctions".
}
\details{
NEEDED
}
