% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_long_pipeline.R
\name{bulk_long_pipeline}
\alias{bulk_long_pipeline}
\title{Title}
\usage{
bulk_long_pipeline(
  annot,
  fastq_dir,
  bam,
  outdir,
  genome_fa,
  minimap2_dir,
  config_file,
  downsample_ratio
)
}
\arguments{
\item{annot}{gene annotations in gff3 file format}

\item{fastq_dir}{the folder containing fastq files, each containing data from one sample}

\item{bam}{aligned bam file (sorted and indexed). Overwrite --infq?}

\item{outdir}{directory to deposite all results in rootdir. Absolute path}

\item{genome_fa}{genome fasta file}

\item{minimap2_dir}{directory containing minimap2, k8 and paftools.js program. k8 and paftools.js are used to convert gff3 to bed12}

\item{config_file}{json configuration files}

\item{downsample_ratio}{downsampling ratio if performing downsampling analysis}
}
\description{
Desc
}
\details{
THIS NEED PROPER TYPES
}
