% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_pipelines.R
\name{bulk_windows_pipeline_setup}
\alias{bulk_windows_pipeline_setup}
\title{FLAMES Windows Bulk Pipeline}
\usage{
bulk_windows_pipeline_setup(
  annot,
  fastq,
  in_bam = NULL,
  outdir,
  genome_fa,
  downsample_ratio = 1,
  config_file
)
}
\arguments{
\item{annot}{gene annotations file in gff3  format}

\item{fastq}{file path to input fastq file}

\item{in_bam}{optional bam file to use instead of fastq files (skips read alignment step)}

\item{outdir}{directory to store all output files.}

\item{genome_fa}{genome fasta file.}

\item{downsample_ratio}{downsampling ratio if performing downsampling analysis.}

\item{config_file}{JSON configuration file. If specified,  \code{config_file} overrides
all configuration parameters}
}
\value{
a list \code{pipeline_variables} with the required variables for execution of later Windows pipeline
steps. File paths required to perform minimap2 alignment are given in pipeline_variables$return_files.
This list should be given as input for \code{windows_pipeline_isoforms} after minimap2 alignment has taken place; \code{windows_pipeline_isoforms} is the
continuation of this pipeline.
}
\description{
An implementation of the FLAMES pipeline designed to run on Windows, or any OS 
without access to minimap2, for read realignment.
This pipeline requires external read alignment, in betwen pipeline calls.
}
\details{
This function, \code{bulk_windows_pipeline_setup} is the first step in the 3 step Windows FLAMES
bulk pipeline, and should be run first, read alignment undertaken,
then \code{windows_pipline_isoforms} should be run, read realignment performed, and finally
\code{windows_pipeline_quantification} should be run.
For each function, besides \code{bulk_windows_pipeline_setup}, a list \code{pipeline_variables} is returned, which
contains the information required to continue the pipeline. This list should be passed into each function, and updated
with the returned list. In the case of \code{bulk_windows_pipeline_setup}, \code{pipeline_variables} is
the list returned. See the vignette 'Vignette for FLAMES bulk on Windows' for more details.
}
