% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_posterior}
\alias{c_posterior}
\title{c posterior distribution}
\usage{
c_posterior(y, X, p_beta, p_c, p_df, inv_link_f, a_c = 1, b_c = 1,
  log = TRUE, method = "ARMS")
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{p_beta}{Regression coefficients}

\item{p_c}{c parameter}

\item{p_df}{Degrees of freedom}

\item{inv_link_f}{Inverse link function}

\item{a_c}{Shape1 for c prior}

\item{b_c}{Shape2 for c prior}

\item{log}{For log-scale values}

\item{method}{"ARMS" or "metropolis"}
}
\value{
Likelihood value
}
\description{
c posterior distribution
}
