% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_basilisk.R
\name{callBasilisk}
\alias{callBasilisk}
\title{Internal utility function for simplifying
calls to basiliskRun using a given basilisk environment}
\usage{
callBasilisk(env_name, FUN, ...)
}
\arguments{
\item{env_name}{the name of the basilisk env (made through BasiliskEnvironment) to execute code within}

\item{FUN}{the function to execute from with the basilisk environment}

\item{...}{extra parameters required by FUN}
}
\value{
the result of `FUN`
}
\description{
Internal utility function for simplifying
calls to basiliskRun using a given basilisk environment
}
