% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.crobit}
\alias{coef.crobit}
\title{Coefficients for crobit class}
\usage{
\method{coef}{crobit}(object, ...)
}
\arguments{
\item{object}{crobit object to restore the coefficients}

\item{...}{Other parameters for coef}
}
\description{
Coefficients for crobit class
}
