% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.FLAMES}
\alias{coef.FLAMES}
\title{Coefficients for FLAMES class}
\usage{
\method{coef}{FLAMES}(object, ...)
}
\arguments{
\item{object}{FLAMES object to restore the coefficients}

\item{...}{Other parameters for coef}
}
\description{
Coefficients for FLAMES class
}
