% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.robit}
\alias{coef.robit}
\title{Coefficients for robit class}
\usage{
\method{coef}{robit}(object, ...)
}
\arguments{
\item{object}{robit object to restore the coefficients}

\item{...}{Other parameters for coef}
}
\description{
Coefficients for robit class
}
