% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tbreg}
\alias{coef.tbreg}
\title{Coefficients for tbreg class}
\usage{
\method{coef}{tbreg}(object, ...)
}
\arguments{
\item{object}{tbreg object to restore the coefficients}

\item{...}{Other parameters for coef}
}
\description{
Coefficients for tbreg class
}
