% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_annotate_plots.R
\name{combine_sce}
\alias{combine_sce}
\title{Combine SCE}
\usage{
combine_sce(
  short_read_large,
  short_read_small,
  short_read_all,
  long_read_sce,
  remove_duplicates = TRUE
)
}
\arguments{
\item{short_read_large}{The SCE object, or path to the HDF5 file, or folder containing the matrix file,
corresponding to the larger short-read sample}

\item{short_read_small}{The SCE object, or path to the HDF5 file, or folder containing the matrix file,
corresponding to the smaller short-read sample}

\item{short_read_all}{(Optional) The combined short-read SCE object. When this is provided, \code{short_read_large} and
\code{short_read_small} will be ignored.}

\item{long_read_sce}{The SCE object of the transcript counts, from the long-read pipelines.}

\item{remove_duplicates}{determines whether cells with duplicated barcodes aer kept in the smaller library (
they are always removed from the larger library)}
}
\value{
A \code{SingleCellExperiment} object, with the transcript counts attached.
}
\description{
Combine long- and short-read SingleCellExperiment objects
}
\details{
Takes the long-read SCE object from the long-read pipeline and attach it as \code{altExpress} of
the short-read SCE object. Cell that are not sampled for long-read sequencing are left with zero expressions.
Cells with duplicated barcodes are removed from the larger library.
Cell barcodes only found in long-read sequencing are removed.
}
\examples{
library(SingleCellExperiment)
a <- SingleCellExperiment(assays = list(counts = matrix(rpois(100, 5), ncol = 10)))
b <- SingleCellExperiment(assays = list(counts = matrix(rpois(100, 5), ncol = 10)))
long_read <- SingleCellExperiment(assays = list(counts = matrix(rpois(100, 5), ncol = 10)))
colData(a)$Barcode <- paste0(1:10, "-1")
colData(b)$Barcode <- paste0(8:17, "-1")
colnames(long_read) <- as.character(2:11)
rownames(a) <- as.character(101:110)
rownames(b) <- as.character(103:112)
rownames(long_read) <- as.character(1001:1010)
combine_sce(short_read_large = a, short_read_small = b, long_read_sce = long_read)

}
