% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CPO_bernoulli}
\alias{CPO_bernoulli}
\title{Conditional Predictive Ordinate for Poisson data}
\usage{
CPO_bernoulli(y, p)
}
\arguments{
\item{y}{Vector of response variable}

\item{p}{Matrix of probabilities for each individual y in each MCMC iteration}
}
\value{
cpo Conditional Predictive Ordinate
}
\description{
Conditional Predictive Ordinate for Poisson data
}
