% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{create_config}
\alias{create_config}
\title{Create Configuration File From Arguments}
\usage{
create_config(
  outdir,
  do_genome_align,
  do_isoform_id,
  do_read_realign,
  do_transcript_quanti,
  gen_raw_isoform,
  has_UMI,
  MAX_DIST,
  MAX_TS_DIST,
  MAX_SPLICE_MATCH_DIST,
  min_fl_exon_len,
  Max_site_per_splice,
  Min_sup_cnt,
  Min_cnt_pct,
  Min_sup_pct,
  strand_specific,
  remove_incomp_reads,
  use_junctions,
  no_flank,
  use_annotation,
  min_tr_coverage,
  min_read_coverage
)
}
\arguments{
\item{outdir}{the destination directory for the configuratio nfile}

\item{do_genome_align}{Boolean. Specifies whether to run the genome alignment step. \code{TRUE} is recommended}

\item{do_isoform_id}{Boolean. Specifies whether to run the isoform identification step. \code{TRUE} is recommended}

\item{do_read_realign}{Boolean. Specifies whether to run the read realignment step. \code{TRUE} is recommended}

\item{do_transcript_quanti}{Boolean. Specifies whether to run the transcript quantification step. \code{TRUE} is recommended}

\item{gen_raw_isoform}{Boolean.}

\item{has_UMI}{Boolean. Specifies if the data contains UMI.}

\item{MAX_DIST}{Maximum distance allowed when merging splicing sites in isoform consensus clustering.}

\item{MAX_TS_DIST}{Maximum distance allowed when merging transcript start/end position in isoform consensus clustering.}

\item{MAX_SPLICE_MATCH_DIST}{Maximum distance allowed when merging splice site called from the data and the reference annotation.}

\item{min_fl_exon_len}{Minimum length for the first exon outside the gene body in reference annotation. This is to correct the alignment artifact}

\item{Max_site_per_splice}{Maximum transcript start/end site combinations allowed per splice chain}

\item{Min_sup_cnt}{Minimum number of read support an isoform decrease this number will significantly increase the number of isoform detected.}

\item{Min_cnt_pct}{Minimum percentage of count for an isoform relative to total count for the same gene.}

\item{Min_sup_pct}{Minimum percentage of count for an splice chain that support a given transcript start/end site combination.}

\item{strand_specific}{1, -1 or 0. 1 indicates if reads are in the same
strand as mRNA, -1 indicates reads are reverse complemented, 0 indicates
reads are not strand specific.}

\item{remove_incomp_reads}{The strenge of truncated isoform filtering. larger number means more stringent filtering.}

\item{use_junctions}{whether to use known splice junctions to help correct the alignment results}

\item{no_flank}{Boolean. for synthetic spike-in data. refer to Minimap2 document for detail}

\item{use_annotation}{Boolean. whether to use reference to help annotate known isoforms}

\item{min_tr_coverage}{Minimum percentage of isoform coverage for a read to be aligned to that isoform}

\item{min_read_coverage}{Minimum percentage of read coverage for a read to be uniquely aligned to that isoform}
}
\value{
the created R list object, which represents the configuration file created by this function.
}
\description{
Create Configuration File From Arguments
}
\details{
Create a list object containing the arguments supplied in a format usable for the FLAMES pipeline. 
Also writes the object to a JSON file, which is located with the prefix 'config_' in the supplied \code{outdir}.
}
\examples{
# create the default configuartion file
\dontrun{
config <- create_config(getwd(), TRUE, TRUE, 
                        TRUE, TRUE,
                        TRUE, FALSE,
                        10, 100, 10,
                        40, 3, 10,
                        0.01, 0.2, 1, 5,
                        TRUE, TRUE,
                        TRUE, 0.75, 0.75)
}
}
