% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{cutadapt}
\alias{cutadapt}
\title{cutadapt wrapper}
\usage{
cutadapt(args)
}
\arguments{
\item{args}{arguments to be passed to cutadapt}
}
\value{
Exit code of cutadapt
}
\description{
trim TSO adaptor with cutadapt
}
\examples{
\dontrun{
 cutadapt("-h")
}
}
