% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d_fullcond}
\alias{d_fullcond}
\title{d full conditional distribution}
\usage{
d_fullcond(y, X, p_beta, p_c, p_d, p_df, inv_link_f, a_d, b_d,
  log = TRUE, method = "ARMS")
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{p_beta}{Regression coefficients}

\item{p_c}{c parameter}

\item{p_d}{d parameter}

\item{p_df}{Degrees of freedom}

\item{inv_link_f}{Inverse link function}

\item{a_d}{Shape1 for d prior}

\item{b_d}{Shape2 for d prior}

\item{log}{For log-scale values}

\item{method}{"ARMS" or "metropolis"}
}
\value{
Likelihood value
}
\description{
d full conditional distribution
}
