% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars.R
\docType{data}
\name{default_config_file}
\alias{default_config_file}
\title{Default Configuration File}
\format{
An object of class \code{character} of length 1.
}
\usage{
default_config_file
}
\description{
file path to the default FLAMES configuration file
}
\keyword{datasets}
