% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_sockey.R
\name{demultiplex_sockeye}
\alias{demultiplex_sockeye}
\title{Demultiplex reads using Sockeye outputs}
\usage{
demultiplex_sockeye(fastq_dir, sockeye_tsv, out_fq)
}
\arguments{
\item{fastq_dir}{The folder containing FASTQ files from Sockeye's output under \code{ingest/chunked_fastqs}.}

\item{sockeye_tsv}{The \code{cell_umi_gene.tsv} file from Sockeye.}

\item{out_fq}{The output FASTQ file.}
}
\value{
returns NULL
}
\description{
Demultiplex reads using the \code{cell_umi_gene.tsv} file from Sockeye.
}
