% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_fullcond}
\alias{df_fullcond}
\title{Degree of freedom full conditional distribution}
\usage{
df_fullcond(y, X, p_beta, p_c, p_d, p_df, p_lambda, inv_link_f,
  log = TRUE, method = "ARMS", const)
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{p_beta}{Regression coefficients}

\item{p_c}{c parameter}

\item{p_d}{d parameter}

\item{p_df}{Degrees of freedom}

\item{p_lambda}{Lambda hyperparameter for p_df}

\item{inv_link_f}{Inverse link function}

\item{log}{For log-scale values}

\item{method}{"ARMS" or "metropolis"}

\item{const}{For numerical problems}
}
\value{
Likelihood value
}
\description{
Degree of freedom full conditional distribution
}
