% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_posterior}
\alias{df_posterior}
\title{Degree of freedom posterior distribution}
\usage{
df_posterior(y, X, p_beta, p_c, p_df, p_lambda, inv_link_f, log = TRUE,
  method = "ARMS", const = 50)
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{p_beta}{Regression coefficients}

\item{p_c}{c parameter}

\item{p_df}{Degrees of freedom}

\item{p_lambda}{Lambda hyperparameter for p_df}

\item{inv_link_f}{Inverse link function}

\item{log}{For log-scale values}

\item{method}{"ARMS" or "metropolis"}

\item{const}{For numerical problems}
}
\value{
Likelihood value
}
\description{
Degree of freedom posterior distribution
}
