% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(y, p)
}
\arguments{
\item{y}{Vector of response variable}

\item{p}{Matrix of probabilities for each individual y in each MCMC iteration}
}
\value{
DIC Deviance Information Criterion measure
}
\description{
Deviance Information Criterion
}
