% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{fake_stranded_gff}
\alias{fake_stranded_gff}
\title{Fake stranded GFF file}
\usage{
fake_stranded_gff(gff_file)
}
\value{
Path to the temporary file with unstranded transcripts converted to '+'.
}
\description{
Check if all the transcript in the annotation is stranded. If not, convert to '+'.
}
\keyword{internal}
