% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{find_bin}
\alias{find_bin}
\title{Find path to a binary
Wrapper for Sys.which to find path to a binary}
\usage{
find_bin(command)
}
\arguments{
\item{command}{character, the command to search for}
}
\value{
character, the path to the command or \code{NA}
}
\description{
This function is a wrapper for \code{base::Sys.which} to find the path
to a command. It also searches within the \code{FLAMES} basilisk conda
environment. This function also replaces "" with \code{NA} in the
output of \code{base::Sys.which} to make it easier to check if the
binary is found.
}
\examples{
find_bin("minimap2")
}
