% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{find_isoform}
\alias{find_isoform}
\title{Find isoform pipeline step}
\usage{
find_isoform(
  gff3,
  genome_bam,
  isoform_gff3,
  tss_tes_stat,
  genomefa,
  transcript_fa,
  downsample_ratio
)
}
\description{
Find isoform pipeline step
}
