% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_measures}
\alias{fit_measures}
\title{LPML, DIC and WAIC measures}
\usage{
fit_measures(y, p, nrep = NULL, nsamp = 100)
}
\arguments{
\item{y}{Vector of response variable}

\item{p}{Matrix of probabilities for each individual y in each MCMC iteration}

\item{nrep}{Number of replications in bootstrap (null to use all data)}

\item{nsamp}{Number of observation in each bootstrap realization}
}
\value{
measures A data.frame with LMPL, DIC and WAIC measures
}
\description{
LPML, DIC and WAIC measures
}
