% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame_generate_dataframe.R
\name{flame_generate_dataframe}
\alias{flame_generate_dataframe}
\title{Flame Database Generator}
\usage{
flame_generate_dataframe(a_list)
}
\arguments{
\item{a_list}{Flame data list}
}
\description{
Parses a list of flame values and then creates a tibble from this information.
}
\examples{
flame_generate_dataframe()
}
