% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame_text_reader.R
\name{flame_text_reader}
\alias{flame_text_reader}
\title{Flame Text Reader}
\usage{
flame_text_reader(directory)
}
\arguments{
\item{directory}{Folder directory that contains the files}
}
\description{
Coverts all TXT files in a folder into a formatted list that works with the database generator. Current requires all files in folder to be TXT files.
Will fix this later.
}
\examples{
flame_text_reader()
}
