% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMES-class.R
\docType{class}
\name{FLAMES-class}
\alias{FLAMES-class}
\alias{FLAMES_class}
\title{An object of 'FLAMES' class}
\description{
An object of 'FLAMES' class
}
\section{Slots}{

\describe{
\item{\code{p}}{A matrix with estimated probability}

\item{\code{beta}}{A matrix with coefficients samples}

\item{\code{c}}{A vector with c parameter sample}

\item{\code{df}}{A vector with df parameter sample}

\item{\code{lambda}}{A vector with lambda parameter sample}

\item{\code{time}}{Time elapsed}

\item{\code{call}}{Function call}

\item{\code{fit_measures}}{DIC, -2*LPML and WAIC measures}
}}

