\name{FlamesR-package}
\alias{FlamesR-package}
\alias{FlamesR}
\docType{package}
\title{
  Full-length transcriptome splicing and mutation analysis.
}
\details{
  This package provides a pipeline for analysis of single cell data, and a pipeline for the analysis of
  bulk long data. It also provides a number of useful functions, used as intermediate functions in these two pipelines,
  which can be used individually for customised analysis.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  [sc_long_pipeline()] for single cell data,
  [bulk_long_pipeline()] for bulk data.
}
