% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_umap.R
\name{generate_umap}
\alias{generate_umap}
\title{Generate UMAP plot}
\usage{
generate_umap(summarizedExperiment, counts = "counts")
}
\arguments{
\item{summarizedExperiment}{the SingleCellExperiment of SummarizedExperiment
object containing the counts to plot}

\item{counts}{character string indicating the name of the assay to generate UMAP visualtization from.
The assay should be accessable by `assay(summarizedExperiemnt, counts)`.}
}
\value{
Modified SummarizedExperiment (or SingleCellExperiment) containing
the UMAP visualisation coordinates in the `reducedDims()` UMAP slot.
}
\description{
Generates the matrix of UMAP visualisation coordinates from a given 
SummarizedExperiment.
The modified SummarizedExperiment object is returned, with UMAP coordinates
stored in the `reducedDims()` slot.
}
\seealso{
[sc_long_pipeline()]

[bulk_long_pipeline()]
}
