% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_pipeline.R
\name{generic_long_pipeline}
\alias{generic_long_pipeline}
\title{useDynLib(FLAMES, .registration = TRUE)}
\usage{
generic_long_pipeline(
  annot,
  fastq,
  in_bam,
  outdir,
  genome_fa,
  minimap2_dir,
  downsample_ratio,
  config_file,
  do_genome_align,
  do_isoform_id = TRUE,
  do_read_realign,
  do_transcript_quanti,
  gen_raw_isoform,
  has_UMI,
  MAX_DIST,
  MAX_TS_DIST,
  MAX_SPLICE_MATCH_DIST,
  min_fl_exon_len,
  Max_site_per_splice,
  Min_sup_cnt,
  Min_cnt_pct,
  Min_sup_pct,
  strand_specific,
  remove_incomp_reads,
  use_junctions,
  no_flank,
  use_annotation,
  min_tr_coverage,
  min_read_coverage
)
}
\description{
useDynLib(FLAMES, .registration = TRUE)
}
