% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars.R
\name{get_default_config_file}
\alias{get_default_config_file}
\title{Default Configuration File}
\usage{
get_default_config_file()
}
\value{
file path to the FLAMES default configuration file
}
\description{
file path to the default FLAMES configuration file
}
\examples{
config <- parse_json_config(get_default_config_file())
print_config(config)
}
