% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{get_gene_blocks}
\alias{get_gene_blocks}
\title{Get Gene Blocks}
\usage{
get_gene_blocks(gene_dict, chr_to_gene, gene_to_transcript)
}
\arguments{
\item{gene_dict}{NEEDED}

\item{chr_to_gene}{NEEDED}

\item{gene_to_transcript}{NEEDED}
}
\value{
NEEDED
}
\description{
NEEDED
}
\examples{
gff3_parse <- parse_gff_tree(system.file("extdata/SIRV_anno.gtf", package="FLAMES"))
gene_dict <- get_gene_flat(gff3_parse$gene_to_transcript, gff3_parse$transcript_to_exon)
chr_to_blocks <- get_gene_blocks(gene_dict, gff3_parse$chr_to_gene, gff3_parse$gene_to_transcript)
}
