% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{get_gene_flat}
\alias{get_gene_flat}
\title{Get Gene Flat}
\usage{
get_gene_flat(gene_to_transcript, transcript_to_exon)
}
\arguments{
\item{gene_to_transcript}{NEEDED}

\item{transcript_to_exon}{NEEDED}
}
\value{
NEEDED
}
\description{
NEEDED
}
\examples{
gff3_parse <- parse_gff_tree(system.file("extdata/SIRV_anno.gtf", package="FlamesR"))
gene_dict <- get_gene_flat(gff3_parse$gene_to_transcript, gff3_parse$transcript_to_exon)
}
