% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_to_fa.R
\name{get_transcript_seq}
\alias{get_transcript_seq}
\title{Title}
\usage{
get_transcript_seq(
  fa_file,
  fa_out_f,
  chr_to_gene,
  transcript_dict,
  gene_to_transcript,
  transcript_to_exon,
  ref_dict = NULL
)
}
\arguments{
\item{name}{desc}
}
\description{
DESC
}
