% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{gff3_to_bed12}
\alias{gff3_to_bed12}
\title{Title}
\usage{
gff3_to_bed12(minimap2_prog_path, gff3_file, bed12_file)
}
\arguments{
\item{minimap2_prog_path}{Directory containing minimap2, k8 and paftools.js. These are
used to convert gff3 to bed12.}

\item{gff3_file}{The gff3_file to convert}

\item{bed12_file}{The filename of the bed12 output file.}
}
\description{
DESC
}
