% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread_functions.R
\name{group_bam2isoform}
\alias{group_bam2isoform}
\title{Group BAM to Isoform}
\usage{
group_bam2isoform(
  bam_in,
  out_gff3,
  out_stat,
  summary_csv,
  chr_to_blocks,
  gene_dict,
  transcript_to_junctions,
  transcript_dict,
  fa_f,
  config,
  downsample_ratio,
  raw_gff3 = NULL
)
}
\arguments{
\item{name}{desc}
}
\description{
DESC
}
