% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_link}
\alias{inv_link}
\title{Inverse link function for several models (cumulative distributions)}
\usage{
inv_link(x, type = "logit", df = 1)
}
\arguments{
\item{x}{Value to evaluate the cumulative distribution}

\item{type}{"logit", "probit", "cauchit", "robit" or "cloglog"}

\item{df}{Degrees of freedom (if type = "robit")}
}
\value{
Cumulative probability in x
}
\description{
Inverse link function for several models (cumulative distributions)
}
