% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{locate_minimap2_dir}
\alias{locate_minimap2_dir}
\title{Locate minimap2}
\usage{
locate_minimap2_dir(minimap2_dir = NULL)
}
\arguments{
\item{minimap2_dir}{User's input for \code{minimap2_dir}}
}
\value{
Path to folder containing minimap2
}
\description{
Locate the folder containing minimap2, stop if minimap2 is not
available in the environment or in the user provided folder. If the user provided
 a path to a file, check if it is minimap2 and return its parent folder.
}
\examples{
mm2_dir <- locate_minimap2_dir()
}
