% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_p}
\alias{make_p}
\title{Calculate p}
\usage{
make_p(X, p_beta, p_c, p_df, inv_link_f)
}
\arguments{
\item{X}{Covariate matrix}

\item{p_beta}{Coefficients}

\item{p_c}{c parameter}

\item{p_df}{Degrees of freedom}

\item{inv_link_f}{Inverse link function}
}
\value{
Likelihood value
}
\description{
Calculate p
}
