% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_bin_arms.R
\name{mcmc_bin_arms}
\alias{mcmc_bin_arms}
\title{MCMC for binary regression - ARMS version}
\usage{
mcmc_bin_arms(y, X, nsim, burnin, lag, inv_link_f, type, sample_c,
  sigma_beta, a_c, b_c, a_lambda, b_lambda, bound_beta, p_c, p_prop,
  p_beta, p_df, p_lambda, const)
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{nsim}{Sample size required for MCMC}

\item{burnin}{Burn in for MCMC}

\item{lag}{Lag for MCMC}

\item{inv_link_f}{Inverse link function}

\item{type}{"logit", "probit", "cauchit", "robit" or "cloglog"}

\item{sample_c}{Should c be sampled?}

\item{sigma_beta}{Variance of beta prior}

\item{a_c}{Shape1 for c prior}

\item{b_c}{Shape2 for c prior}

\item{a_lambda}{Inferior limit for lambda}

\item{b_lambda}{Superior limit for lambda}

\item{bound_beta}{Limits to sample beta}

\item{p_c}{To restore c}

\item{p_prop}{To restore p}

\item{p_beta}{To restore beta}

\item{p_df}{To restore df}

\item{p_lambda}{To restore lambda}

\item{const}{A constant to help on sampling degrees of freedom \eqn{\tilde{df} = df/c}}
}
\value{
Chains of all parameters
}
\description{
Perform MCMC with ARMS algorithm
}
