% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_bin_metropolis.R
\name{mcmc_bin_metropolis}
\alias{mcmc_bin_metropolis}
\title{MCMC for binary regression - Metropolis version}
\usage{
mcmc_bin_metropolis(y, X, nsim, burnin, lag, inv_link_f, type, sample_c,
  sample_d, sigma_beta, a_c, b_c, a_d, b_d, a_lambda, b_lambda, var_df,
  var_c, var_d, var_lambda, p_c, p_d, p_prop, p_beta, p_df, p_lambda,
  const)
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{nsim}{Sample size required for MCMC}

\item{burnin}{Burn in for MCMC}

\item{lag}{Lag for MCMC}

\item{inv_link_f}{Inverse link function}

\item{type}{"logit", "probit", "cauchit", "robit" or "cloglog"}

\item{sample_c}{Should c be sampled?}

\item{sample_d}{Should c be sampled?}

\item{sigma_beta}{Variance of beta prior}

\item{a_c}{Shape1 for c prior}

\item{b_c}{Shape2 for c prior}

\item{a_d}{Shape1 for d prior}

\item{b_d}{Shape2 for d prior}

\item{a_lambda}{Inferior limit for lambda}

\item{b_lambda}{Superior limit for lambda}

\item{var_df}{Variance to sample 1-exp(-df/const)}

\item{var_c}{Variance to sample from c (if sample_d = TRUE) otherwise log(c/(1-c))}

\item{var_d}{Variance to sample d (if sample_c = TRUE) otherwise log(c/(1-c))}

\item{var_lambda}{Variance to sample lambda}

\item{p_c}{To restore c}

\item{p_d}{To restore d}

\item{p_prop}{To restore p}

\item{p_beta}{To restore beta}

\item{p_df}{To restore df}

\item{p_lambda}{To restore lambda}

\item{const}{A constant to help on sampling degrees of freedom \eqn{\tilde{df} = df/c}}
}
\value{
Chains of all parameters
}
\description{
Perform MCMC with Metropolis Hastings algorithm
}
