% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bulk.R
\name{merge_bulk_fastq}
\alias{merge_bulk_fastq}
\title{Merge FASTQ}
\usage{
merge_bulk_fastq(fastq_dir, out_fastq)
}
\arguments{
\item{fastq_dir}{The folder containing fastq files to merge}

\item{out_fastq}{A file which will be created to store all fastq entries. Overwrites existing files}
}
\value{
file path to the created merged fastq file `out_fastq`
}
\description{
Merges all fastq files in the given folder into a single file. Used to create a pseudobulk fastq file,
with added fake barcodes to differentiate between source files.
}
\examples{
out_fastq <- merge_bulk_fastq(system.file("extdata/fastq", package="FLAMES"), anno_csv=tempfile(), out_fastq=tempfile())
}
