% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_bulk_fastq_cpp}
\alias{merge_bulk_fastq_cpp}
\title{Merge Bulk Fastq Files}
\usage{
merge_bulk_fastq_cpp(fastq_files, out_fastq)
}
\arguments{
\item{fastq_files}{a string vector of the fastq file paths to merge}

\item{out_fastq}{the fastq file path of the output file}
}
\value{
returns NULL
}
\description{
Merge all fastq files into a single fastq
This function also inserts the original fastq file name into the header line of each read.
}
