% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{minimap2_align}
\alias{minimap2_align}
\title{Minimap2 Align to Genome}
\usage{
minimap2_align(
  minimap2_prog_path = NULL,
  fa_file,
  fq_in,
  bam_out,
  no_flank = FALSE,
  bed12_junc = NULL
)
}
\arguments{
\item{minimap2_prog_path}{Absolute path to the directory containing minimap2}

\item{fa_file}{Fasta file used as a reference database for alignment}

\item{fq_in}{Fastq file used as a query sequence file}

\item{bam_out}{Output BAM file}

\item{no_flank}{Boolean; used if studying SIRV, to let minimap2 ignore additional bases}

\item{bed12_junc}{Gene annotations in BED12 format. If specified, minmap2 prefers splicing in annotations.}
}
\value{
file path to the given output BAM file, \code{bam_out}
}
\description{
Uses minimap2 to align sequences agains a reference databse.
}
\details{
NEEDED. What does minimap2 do, and why is it used in FLAMES?
}
