% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{minimap2_realign}
\alias{minimap2_realign}
\title{Minimap2 re-align reads to transcriptome}
\usage{
minimap2_realign(
  config,
  fa_file,
  fq_in,
  outdir,
  minimap2_dir,
  prefix = NULL,
  threads = NULL
)
}
\arguments{
\item{config}{Parsed list of FLAMES config file}

\item{fa_file}{Path to the transcript assembly (\code{.fasta}) file}

\item{fq_in}{File path to the fastq file used as a query sequence file}

\item{outdir}{Output folder}

\item{minimap2_dir}{Path to the directory containing minimap2}

\item{prefix}{String, the prefix (e.g. sample name) for the outputted BAM file}

\item{threads}{Integer, threads for minimap2 to use, see minimap2 documentation for details,
FLAMES will try to detect cores if this parameter is not provided.}
}
\value{
NULL, BAM file is saved to \code{outdir}
}
\description{
Uses minimap2 to re-align reads to transcriptome
}
\seealso{
[minimap2_align()]
}
