% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{minimap2_tr_align}
\alias{minimap2_tr_align}
\title{Minimap2 Align to Transcript}
\usage{
minimap2_tr_align(mm2_prog_path, fa_file, fq_in, bam_out)
}
\arguments{
\item{mm2_prog_path}{Absolute path to the directory containing minimap2}

\item{fa_file}{Input fasta file used as a reference database}

\item{fq_in}{Input fastq used as a query sequence file}

\item{bam_out}{Output BAM file, containing aligned sequences}
}
\value{
file path to the create bam file `bam_out`
}
\description{
Uses minimap2 to align to transcript. 
Uses options "-ax map-ont -p 0.9 --end-bonus 10 -N 3 -t 12 \code{fa_file} \code{fq_in} | samtools view -bS -@ 4 -m 2G -o \code{bam_out}"
}
