% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gene_anno.R
\name{parse_gff_tree}
\alias{parse_gff_tree}
\title{Parse Gff3 file}
\usage{
parse_gff_tree(gff_file)
}
\arguments{
\item{gff_file}{the gff3 file to parse}
}
\value{
a named list of elements:
"chr_to_gene", "transcript_dict", "gene_to_transcript", "transcript_to_exon"
}
\description{
Parse a Gff3 file into 3 components: chromasome to ene, transcript dictionary, gene to transcript dictionary
and a transcript to exon dictionary.
These components are returned in a named list.
}
