% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R, R/parse_json_config.R
\name{parse_json_config}
\alias{parse_json_config}
\title{Title}
\usage{
parse_json_config(json_config)

parse_json_config(json_config)
}
\arguments{
\item{name}{desc}
}
\description{
DESC

DESC
}
