% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tr.R
\name{parse_realigned_bam}
\alias{parse_realigned_bam}
\title{Parse Realigned Bam}
\usage{
parse_realigned_bam(
  bam_in,
  fa_idx_f,
  min_sup_reads,
  min_tr_coverage,
  min_read_coverage,
  ...
)
}
\description{
Parse Realigned Bam
}
\examples{

realign_bam <- system.file("extdata/align2genome.bam", package="FLAMES")
transcript_fa <- system.file("extdata/transcript_assembly.fa.fai", package="FLAMES")
\dontrun{
parse_realign <- parse_realigned_bam(realign_bam, transcript_fa, 10, 0.75, 0.75)
}
}
