% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.crobit}
\alias{plot.crobit}
\title{Plot for crobit class}
\usage{
\method{plot}{crobit}(x, bty = "l", ask = T, ...)
}
\arguments{
\item{x}{crobit object to plot}

\item{bty}{Box style}

\item{ask}{To ask for next plot}

\item{...}{Other parameters for plot}
}
\description{
Plot for crobit class
}
