% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.FLAMES}
\alias{plot.FLAMES}
\title{Plot for FLAMES class}
\usage{
\method{plot}{FLAMES}(x, bty = "l", ask = T, ...)
}
\arguments{
\item{x}{FLAMES object to plot}

\item{bty}{Box style}

\item{ask}{To ask for next plot}

\item{...}{Other parameters for plot}
}
\description{
Plot for FLAMES class
}
