% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.robit}
\alias{plot.robit}
\title{Plot for robit class}
\usage{
\method{plot}{robit}(x, bty = "l", ask = T, ...)
}
\arguments{
\item{x}{robit object to plot}

\item{bty}{Box style}

\item{ask}{To ask for next plot}

\item{...}{Other parameters for plot}
}
\description{
Plot for robit class
}
