% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tbreg}
\alias{plot.tbreg}
\title{Plot for tbreg class}
\usage{
\method{plot}{tbreg}(x, bty = "l", ask = T, ...)
}
\arguments{
\item{x}{tbreg object to plot}

\item{bty}{Box style}

\item{ask}{To ask for next plot}

\item{...}{Other parameters for plot}
}
\description{
Plot for tbreg class
}
