% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_barcode.R
\name{plot_demultiplex}
\alias{plot_demultiplex}
\title{Plot Cell Barcode demultiplex statistics}
\usage{
plot_demultiplex(outdir, stats_file)
}
\arguments{
\item{outdir}{folder containing the \code{matched_barcode_stat} file, or
\code{matched_barcode_stat.SAMPLE} files. Ignored if \code{stats_file} is provided.}

\item{stats_file}{\code{matched_barcode_stat} file(s) from which the statistics to be plotted.}
}
\value{
a \code{ggplot} object of the barcode plot
}
\description{
produce a barplot of cell barcode demultiplex statistics
}
\examples{
outdir <- tempfile()
dir.create(outdir)
bc_allow <- file.path(outdir, "bc_allow.tsv")
R.utils::gunzip(filename = system.file("extdata/bc_allow.tsv.gz", package = "FLAMES"), destname = bc_allow, remove = FALSE)
find_barcode(
  fastq = system.file("extdata/fastq", package = "FLAMES"),
  stats_out = file.path(outdir, "bc_stat"),
  reads_out = file.path(outdir, "demultiplexed.fq.gz"),
  barcodes_file = bc_allow
)
plot_demultiplex(stats_file = file.path(outdir, "bc_stat"))
}
