% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_spatial_isoform}
\alias{plot_spatial_isoform}
\title{Plot spatial pie chart of isoforms}
\usage{
plot_spatial_isoform(spe, isoforms, assay_type = "counts")
}
\arguments{
\item{spe}{The SpatialExperiment object.}

\item{isoforms}{The isoforms to plot.}

\item{assay_type}{The assay that contains the given features. E.g. 'counts', 'logcounts'.}
}
\value{
A ggplot object.
}
\description{
This function plots a spatial pie chart for given features.
}
