% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_spatial_pie}
\alias{plot_spatial_pie}
\title{Plot spatial pie chart}
\usage{
plot_spatial_pie(spe, features, assay_type = "counts")
}
\arguments{
\item{spe}{The SpatialExperiment object.}

\item{features}{The features to plot.}

\item{assay_type}{The assay that contains the given features.}
}
\value{
A ggplot object.
}
\description{
This function plots a spatial pie chart for given features.
}
