% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.crobit}
\alias{print.crobit}
\title{Print for crobit class}
\usage{
\method{print}{crobit}(x, ...)
}
\arguments{
\item{x}{crobit object to print}

\item{...}{Other parameters for print}
}
\description{
Print for crobit class
}
