% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.FLAMES}
\alias{print.FLAMES}
\title{Print for FLAMES class}
\usage{
\method{print}{FLAMES}(x, ...)
}
\arguments{
\item{x}{FLAMES object to print}

\item{...}{Other parameters for print}
}
\description{
Print for FLAMES class
}
