% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robit}
\alias{print.robit}
\title{Print for robit class}
\usage{
\method{print}{robit}(x, ...)
}
\arguments{
\item{x}{robit object to print}

\item{...}{Other parameters for print}
}
\description{
Print for robit class
}
