% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tbreg}
\alias{print.tbreg}
\title{Print for tbreg class}
\usage{
\method{print}{tbreg}(x, ...)
}
\arguments{
\item{x}{tbreg object to print}

\item{...}{Other parameters for print}
}
\description{
Print for tbreg class
}
