% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{print_config}
\alias{print_config}
\title{Print Configuration File}
\usage{
print_config(config)
}
\arguments{
\item{config}{List; the configuration list to print.}
}
\description{
Print Configuration File
}
\details{
Print the configuration file, represented as a named list used for the Flames pipeline.
}
