% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_mutations.R
\name{relative_mutation_positions_single}
\alias{relative_mutation_positions_single}
\title{Relative mutation positions within the gene body}
\usage{
relative_mutation_positions_single(
  mutations,
  annotation_grange,
  verbose = TRUE
)
}
\arguments{
\item{mutations}{either the tibble output from \code{find_variants} or a GRanges object.
Make sure to filter it for only the gene of interest.}

\item{annotation_grange}{GRanges: the gene annotation. Must have the following types: "gene" and "exon".}

\item{verbose}{logical(1): whether to print messages.}
}
\value{
A numeric vector of relative positions of each mutation within the gene body. Ranging from
0 (start of the gene) to 1 (end of the gene).
}
\description{
Given a set of mutations and a gene annotation, calculate the relative position of each mutation
within the gene body. The gene annotation must have the following types: "gene" and "exon".
The gene annotation must be for one gene only. The mutations must be within the gene region.
The function will merge overlapping exons and calculate the relative position of each mutation
within the gene body, excluding intronic regions.
}
