% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{samtools_as_bam}
\alias{samtools_as_bam}
\title{Samtools Convert SAM to BAM}
\usage{
samtools_as_bam(sam_in, bam_out)
}
\arguments{
\item{sam_in}{input sam file}

\item{bam_out}{output bam file}
}
\value{
the path to the output file, given as \code{bam_out}
}
\description{
Converts a .sam file to a .bam file using Rsamtools
}
