% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_annotate_umap.R
\name{sc_annotate_umap}
\alias{sc_annotate_umap}
\title{FLAMES Annotated Plottings}
\usage{
sc_annotate_umap(
  sce_20,
  sce_80,
  path,
  gene,
  n_isoforms = 4,
  n_pcs = 20,
  dup_bc = NULL,
  sce_all = NULL,
  return_sce_all = T
)
}
\arguments{
\item{sce_20}{The \code{SingleCellExperiment} object containing gene counts assay for the smaller library.}

\item{sce_80}{The \code{SingleCellExperiment} object containing gene counts assay for the larger library.}

\item{path}{The path to the folder containing outputs of \code{sc_long_pipeline}.}

\item{gene}{The gene symbol of interest.}

\item{n_isoforms}{The number of expressed isoforms to keep.}

\item{dup_bc}{Cell barcodes found both in the larger and smaller library, will be used to filter cells in the long-read
data. (Filtering long-read data will be implemented in the main pipeline soon)}

\item{sce_all}{The \code{SingleCellExperiment} object containing gene counts assay for the combined library.}

\item{return_sce_all}{Whether to return the processed \code{SingleCellExperiment} object.}
}
\value{
a list containing the combined UMAP, the isoform exon alignments and the UMAP with isoform expression levels.
}
\description{
Plot isoform exons alignments for a given gene, along with UMAP showing expression levels.
}
\details{
This function takes the short-read data (as \code{SingleCellExperiment} objects) from both the smaller
and the larger libraries to generate a combined UMAP, the expression levels of isoforms (using long read data)
are then overlayed on top of the UMAP. SNN inference based on gene counts were performed to impute isoform expression
for cells in the smaller library.
}
