% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_DTU_analysis.R
\name{sc_DTU_analysis}
\alias{sc_DTU_analysis}
\title{FLAMES Differential Transcript Usage Analysis}
\usage{
sc_DTU_analysis(sce, path)
}
\arguments{
\item{sce}{The \code{SingleCellExperiment} object from \code{sc_long_pipeline}, an additional \code{cluster_annotation.csv} 
file is required under the output folder of the SCE object.}

\item{path}{The path to the otput folder of \code{sc_long_pipeline} 
the folder needs to contain:
\itemize{
 \item{transcript_count.csv.gz}{ - the transcript count matrix }
 \item{isoform_FSM_annotation.csv}{ - the full splice match annotation file }
 \item{cluster_annotation.csv}{ - cluster annotation file }
}}
}
\value{
a \code{data.frame} containing the following columns:
\itemize{
 \item{gene_name}{ - differentially transcribed genes }
 \item{X_value}{ - the X value for the DTU gene}
 \item{df}{ - degrees of freedom of the approximate chi-squared distribution of the test statistic }
 \item{DTU_tr}{ - the transcript_id with the highest squared residuals}
 \item{DTU_group}{ - the cell group with the highest squared residuals}
 \item{p_value}{ - the p-value for the test}
 \item{adj_p}{ - the adjusted p-value (by Benjamini–Hochberg correction)}
}
The table is sorted by decreasing P-values.
}
\description{
Chi-square based differential transcription usage analysis. This variant is meant for single cell data.
Takes the \code{SingleCellExperiment} object from \code{sc_long_pipeline} as input.
Alternatively, the path to the output folder could be provided instead of the SCE object.
A cluster annotation file \code{cluster_annotation.csv} is required, please provide this file under the
output folder of \code{sc_long_pipeline}.
}
\details{
This function will search for genes that have at least two isoforms, each with more than 15 UMI counts. 
For each gene, the per cell transcript counts were merged by group to generate pseudo bulk samples. 
Grouping is specified by the \code{cluster_annotation.csv} file. 
The top 2 highly expressed transcripts for each group were selected and a UMI count matrix where 
the rows are selected transcripts and columns are groups was used as input to a chi-square test of independence (chisq.test). 
Adjusted P-values were calculated by Benjamini–Hochberg correction.
}
