% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_annotate_plots.R
\name{sc_heatmap_expression}
\alias{sc_heatmap_expression}
\title{FLAMES heetmap plots}
\usage{
sc_heatmap_expression(
  gene,
  multiAssay,
  impute = FALSE,
  n_isoforms = 4,
  transcript_ids,
  n_pcs = 40,
  isoform_legend_width = 7,
  col_low = "#313695",
  col_mid = "#FFFFBF",
  col_high = "#A50026",
  color_quantile = 0.95
)
}
\arguments{
\item{gene}{The gene symbol of interest.}

\item{multiAssay}{The \code{MultiAssayExperiment} object from \code{combine_sce()}.}

\item{impute}{Whether to impute expression levels for cells without transcript counts}

\item{n_isoforms}{The number of expressed isoforms to keep.}

\item{transcript_ids}{specify the transcript ids instead of selecting the top \code{n_isoforms}}

\item{n_pcs}{The number of principal components to generate.}

\item{isoform_legend_width}{The width of isoform legends in heatmaps, in \code{cm}.}

\item{col_low}{Color for cells with low expression levels in UMAPs.}

\item{col_mid}{Color for cells with intermediate expression levels in UMAPs.}

\item{col_high}{Color for cells with high expression levels in UMAPs.}

\item{color_quantile}{The lower and upper expression quantile to be displayed bewteen \code{col_low} and \code{col_high}, e.g. with \code{color_quantile = 0.95}, cells with expressions higher than 95\% of other cells will all be shown in \code{col_high}, and cells with expression lower than 95\% of other cells will all be shown in \code{col_low}.}
}
\value{
a \code{ggplot} object of the heatmap
}
\description{
Plot expression heatmap of top n isoforms of a gene
}
\details{
This function takes the combined \code{MultiAssayExperiment} object from
\code{combine_sce} and plots an expression heatmap with the isoform alignment visualisations.
}
\examples{
source(system.file("examples/scmixology1.R", package = "FLAMES"))
scm_lib80 <- scuttle::addPerCellQC(scm_lib80)
scm_lib20 <- scuttle::addPerCellQC(scm_lib20)
qc_80 <- scuttle::quickPerCellQC(colData(scm_lib80))
qc_20 <- scuttle::quickPerCellQC(colData(scm_lib20))
qc_80$discard[scm_lib80$sum < 10000] <- TRUE
qc_20$discard[scm_lib20$sum < 20000] <- TRUE

combined_sce <- combine_sce(
    short_read_large = scm_lib80[,!qc_80$discard],
    short_read_small = scm_lib20[,!qc_20$discard],
    long_read_sce = scm_lib20_transcripts,
    remove_duplicates = FALSE)

sc_heatmap_expression(gene = "ENSG00000108107", multiAssay = combined_sce)

}
