% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_long_pipeline.R
\name{sc_long_pipeline}
\alias{sc_long_pipeline}
\title{Pipeline for Single Cell Data}
\usage{
sc_long_pipeline(
  annot,
  infq,
  in_bam = NULL,
  outdir,
  genome_fa,
  minimap2_dir = NULL,
  downsample_ratio = 1,
  config_file = NULL,
  do_genome_align = TRUE,
  do_isoform_id = TRUE,
  do_read_realign = TRUE,
  do_transcript_quanti = TRUE,
  gen_raw_isoform = TRUE,
  has_UMI = FALSE,
  MAX_DIST = 10,
  MAX_TS_DIST = 100,
  MAX_SPLICE_MATCH_DIST = 10,
  min_fl_exon_len = 40,
  Max_site_per_splice = 3,
  Min_sup_cnt = 10,
  Min_cnt_pct = 0.01,
  Min_sup_pct = 0.2,
  strand_specific = 1,
  remove_incomp_reads = 5,
  use_junctions = TRUE,
  no_flank = TRUE,
  use_annotation = TRUE,
  min_tr_coverage = 0.75,
  min_read_coverage = 0.75
)
}
\arguments{
\item{annot}{gene annotations file in gff3  format}

\item{in_bam}{aligned bam file (sorted and indexed). If supplied, this overwrites
\code{fastq_dir} and skips the first alignment step}

\item{outdir}{directory to store all output files.}

\item{genome_fa}{genome fasta file.}

\item{minimap2_dir}{directory containing minimap2, k8 and paftools.js program.
k8 and paftools.js are used to convert gff3 to bed12.}

\item{downsample_ratio}{downsampling ratio if performing downsampling analysis.}

\item{config_file}{JSON configuration file. If specified, \code{config_file} overrides
all configuration parameters}

\item{do_genome_align}{Boolean. Specifies whether to run the genome alignment step. \code{TRUE} is recommended}

\item{do_isoform_id}{Boolean. Specifies whether to run the isoform identification step. \code{TRUE} is recommended}

\item{do_read_realign}{Boolean. Specifies whether to run the read realignment step. \code{TRUE} is recommended}

\item{do_transcript_quanti}{Boolean. Specifies whether to run the transcript quantification step. \code{TRUE} is recommended}

\item{has_UMI}{Boolean. Speficies if each gene as a UMI.}

\item{MAX_DIST}{Numeric}

\item{MAX_TS_DIST}{Numeric.}

\item{MAX_SPLICE_MATCH_DIST}{Numeric.}

\item{min_fl_exon_len}{Numeric.}

\item{Max_site_per_splice}{Numeric.}

\item{Min_sup_cnt}{Numeric.}

\item{Min_cnt_pct}{Numeric.}

\item{Min_sup_pct}{Numeric.}

\item{remove_incomp_reads}{Numeric.}

\item{use_junctions}{Boolean.}

\item{no_flank}{Boolean.}

\item{use_annotation}{Boolean.}

\item{min_tr_coverage}{Numeric.}

\item{min_read_coverage}{Numeric.}
}
\description{
DESC
}
