% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_annotate_plots.R
\name{sc_reduce_dims}
\alias{sc_reduce_dims}
\title{runPCA and runUMAP wrapper}
\usage{
sc_reduce_dims(multiAssay, n_pcs = 40, n_hvgs = 2000)
}
\arguments{
\item{multiAssay}{The \code{MultiAssayExperiment} object from \code{combine_sce()}.}

\item{n_pcs}{The number of principal components to generate.}

\item{n_hvgs}{The number of variable genes to use for running PCA.}
}
\value{
the \code{MultiAssayExperiment} with reduced dimensions
}
\description{
runPCA and runUMAP wrapper for combined SCE object from \code{combine_sce}
}
\details{
This function takes the combined \code{MultiAssayExperiment} object from
\code{combine_sce} and run \code{scater::runUMAP} and / or \code{scran::fixedPCA}
}
\examples{
combined_sce <- combine_sce(
    short_read_large = scmixology_lib90,
    short_read_small = scmixology_lib10,
    long_read_sce = scmixology_lib10_transcripts,
    remove_duplicates = FALSE)
sc_reduce_dims(multiAssay = combined_sce)

}
