% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scmixology_lib10_transcripts}
\alias{scmixology_lib10_transcripts}
\title{scMixology long-read transcript counts - sample 2}
\format{
## `scmixology_lib10_transcripts`
A \code{SingleCellExperiment} with 7,240 rows and 60 columns:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE154869>
}
\usage{
scmixology_lib10_transcripts
}
\description{
long-read transcript counts from long and short-read single 
cell RNA-seq profiling of human  lung adenocarcinoma cell 
lines using 10X version 2 chemstry. See
Tian, L. et al. Comprehensive characterization of single-cell
full-length isoforms in human and mouse with long-read sequencing.
Genome Biology 22, 310 (2021).
}
\keyword{datasets}
