% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.crobit}
\alias{summary.crobit}
\title{Summary for crobit class (coda is necessary)}
\usage{
\method{summary}{crobit}(object, HPD = T, ...)
}
\arguments{
\item{object}{crobit object to summarise}

\item{HPD}{HPD intervals or quantiles (TRUE or FALSE)}

\item{...}{Other parameters for summary}
}
\description{
Summary for crobit class (coda is necessary)
}
