% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.FLAMES}
\alias{summary.FLAMES}
\title{Summary for FLAMES class (coda is necessary)}
\usage{
\method{summary}{FLAMES}(object, HPD = T, ...)
}
\arguments{
\item{object}{FLAMES object to summarise}

\item{HPD}{HPD intervals or quantiles (TRUE or FALSE)}

\item{...}{Other parameters for summary}
}
\description{
Summary for FLAMES class (coda is necessary)
}
