% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.robit}
\alias{summary.robit}
\title{Summary for robit class (coda is necessary)}
\usage{
\method{summary}{robit}(object, HPD = T, ...)
}
\arguments{
\item{object}{robit object to summarise}

\item{HPD}{HPD intervals or quantiles (TRUE or FALSE)}

\item{...}{Other parameters for summary}
}
\description{
Summary for robit class (coda is necessary)
}
