% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tbreg}
\alias{summary.tbreg}
\title{Summary for tbreg class (coda is necessary)}
\usage{
\method{summary}{tbreg}(object, HPD = T, ...)
}
\arguments{
\item{object}{tbreg object to summarise}

\item{HPD}{HPD intervals or quantiles (TRUE or FALSE)}

\item{...}{Other parameters for summary}
}
\description{
Summary for tbreg class (coda is necessary)
}
