% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{sys_which}
\alias{sys_which}
\title{Sys.which wrapper
Wrapper for Sys.which that replaces "" with NA}
\usage{
sys_which(command)
}
\arguments{
\item{command}{character, the command to search for}
}
\value{
character, the path to the command or \code{NA}
}
\description{
The \code{base::Sys.which} function returns "" if the command is not found
on some systems and \code{NA} on others. This wrapper replaces "" with
\code{NA}
}
\examples{
sys_which("minimap2")
}
