% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbreg-class.R
\docType{class}
\name{tbreg-class}
\alias{tbreg-class}
\alias{tbreg_class}
\title{An object of 'tbreg' class}
\description{
An object of 'tbreg' class
}
\section{Slots}{

\describe{
\item{\code{p}}{A matrix with estimated probability}

\item{\code{beta}}{A matrix with coefficients samples}

\item{\code{c}}{A vector with c parameter sample}

\item{\code{df}}{A vector with df parameter sample}

\item{\code{lambda}}{A vector with lambda parameter sample}

\item{\code{time}}{Time elapsed}

\item{\code{call}}{Function call}
}}

