% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WAIC}
\alias{WAIC}
\title{Widely Applicable Information Criterion}
\usage{
WAIC(y, p)
}
\arguments{
\item{y}{Vector of response variable}

\item{p}{Matrix of probabilities for each individual y in each MCMC iteration}
}
\value{
WAIC_1 Widely Applicable Information Criterion measure
}
\description{
Widely Applicable Information Criterion
}
