% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_pipelines.R
\name{windows_pipeline_setup}
\alias{windows_pipeline_setup}
\title{Hidden generic pipeline setup function}
\usage{
windows_pipeline_setup(
  annot,
  fastq,
  in_bam = NULL,
  outdir,
  genome_fa,
  downsample_ratio = 1,
  config_file,
  bc_file = NULL,
  bulk = FALSE
)
}
\description{
Hidden generic pipeline setup function
}
