% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{write_config}
\alias{write_config}
\title{Write Configuration Dictionary to File}
\usage{
write_config(config, config_file)
}
\arguments{
\item{config}{List; the configuration list to write to file.}

\item{config_file}{the file to output \code{config} to. Should be .json extension}
}
\value{
returns NULL
}
\description{
Write Configuration Dictionary to File
}
\details{
Write the configuration file, represented as a named list used for the Flames pipeline.
}
