% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLAZE_demultiplexing.R
\name{blaze}
\alias{blaze}
\title{BLAZE Assign reads to cell barcodes.}
\usage{
blaze(blaze_config, fq_in)
}
\arguments{
\item{blaze_config}{List, additional BLAZE configuration parameters}

\item{fq_in}{File path to the fastq file used as a query sequence file}
}
\value{
a \code{data.frame} summarising the reads aligned
}
\description{
Uses BLAZE to generate barcode list and assign reads to cell barcodes. 
Uses default options for BLAZE, see BLAZE documentation for details (https://github.com/shimlab/BLAZE).
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
bc_list_10x_url <- 'https://github.com/shimlab/BLAZE/blob/main/10X_bc/3M-february-2018.zip'
bc_list_10x <- bfc[[names(BiocFileCache::bfcadd(bfc, 'bc_list_10x', bc_list_10x_url))]]
fastq1_url <- 'https://raw.githubusercontent.com/shimlab/BLAZE/main/test/data/FAR20033_pass_51e510db_100.fastq'
fastq1 <- bfc[[names(BiocFileCache::bfcadd(bfc, 'Fastq1', fastq1_url))]]
outdir <- tempfile()
dir.create(outdir)
config = jsonlite::fromJSON(system.file('extdata/blaze_flames.json', package = 'FLAMES'))
config$blaze_parameters['output-prefix'] <- outdir
blaze(config$blaze_parameters, fastq1)
}
