% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLAZE_demultiplexing.R
\name{blaze}
\alias{blaze}
\title{BLAZE Assign reads to cell barcodes.}
\usage{
blaze(expect_cells, fq_in, ...)
}
\arguments{
\item{expect_cells}{Integer, expected number of cells. Note: this could be just a rough estimate. E.g., the targeted number of cells.}

\item{fq_in}{File path to the fastq file used as a query sequence file}

\item{...}{Additional BLAZE configuration parameters. E.g., setting
`'output-prefix'='some_prefix'` is equivalent to specifying `--output-prefix some_prefix` in BLAZE; Similarly,
`overwrite=TRUE` is equivalent to switch on the `--overwrite` option. Note that the specified parameters will
override the parameters specified in the configuration file. All available options can be found at https://github.com/shimlab/BLAZE.}
}
\value{
A \code{data.frame} summarising the reads aligned. Other outputs are written to disk. 
The details of the output files can be found at https://github.com/shimlab/BLAZE.
}
\description{
Uses BLAZE to generate barcode list and assign reads to cell barcodes.
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
fastq1_url <- 'https://raw.githubusercontent.com/shimlab/BLAZE/main/test/data/FAR20033_pass_51e510db_100.fastq'
fastq1 <- bfc[[names(BiocFileCache::bfcadd(bfc, 'Fastq1', fastq1_url))]]
outdir <- tempfile()
dir.create(outdir)
blaze(expect_cells=10, fastq1, overwrite=TRUE)

}
