% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{blocks_to_junctions}
\alias{blocks_to_junctions}
\title{Blocks To Junctions}
\usage{
blocks_to_junctions(block)
}
\arguments{
\item{block}{NEEDED}
}
\value{
The converted junctions named list
}
\description{
Converts \code{block} to a named list containing the keys "left", "right"
and "junctions".
}
\details{
NEEDED
}
\examples{
# parse the gff file before creating junctions from the gene blocks
gff3_parse <- parse_gff_tree(system.file("extdata/SIRV_anno.gtf", package="FLAMES"))
transcript_junction = blocks_to_junctions(gff3_parse$transcript_to_exon[[1]])
}
