% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{create_config}
\alias{create_config}
\title{Create Configuration File From Arguments}
\usage{
create_config(
  outdir,
  do_genome_align,
  do_isoform_id,
  do_read_realign,
  do_transcript_quanti,
  gen_raw_isoform,
  has_UMI,
  MAX_DIST,
  MAX_TS_DIST,
  MAX_SPLICE_MATCH_DIST,
  min_fl_exon_len,
  Max_site_per_splice,
  Min_sup_cnt,
  Min_cnt_pct,
  Min_sup_pct,
  strand_specific,
  remove_incomp_reads,
  use_junctions,
  no_flank,
  use_annotation,
  min_tr_coverage,
  min_read_coverage
)
}
\arguments{
\item{outdir}{the destination directory for the configuratio nfile}

\item{do_genome_align}{NEEDED}

\item{do_isoform_id}{NEEDED}

\item{do_read_realign}{NEEDED}

\item{do_transcript_quanti}{NEEDED}

\item{gen_raw_isoform}{NEEDED}

\item{has_UMI}{NEEDED}

\item{MAX_DIST}{NEEDED}

\item{MAX_TS_DIST}{NEEDED}

\item{MAX_SPLICE_MATCH_DIST}{NEEDED}

\item{min_fl_exon_len}{NEEDED}

\item{Max_site_per_splice}{NEEDED}

\item{Min_sup_cnt}{NEEDED}

\item{Min_cnt_pct}{NEEDED}

\item{Min_sup_pct}{NEEDED}

\item{strand_specific}{NEEDED}

\item{remove_incomp_reads}{NEEDED}

\item{use_junctions}{NEEDED}

\item{no_flank}{NEEDED}

\item{use_annotation}{NEEDED}

\item{min_tr_coverage}{NEEDED}

\item{min_read_coverage}{NEEDED}
}
\value{
the created R list object, which represents the configuration file created by this function.
}
\description{
Create Configuration File From Arguments
}
\details{
Create a list object containing the arguments supplied in a format usable for the FLAMES pipeline. 
Also writes the object to a JSON file, which is located with the prefix 'config_' in the supplied \code{outdir}.
}
\examples{
# create the default configuartion file
config <- create_config(getwd(), TRUE, TRUE, 
                        TRUE, TRUE,
                        TRUE, FALSE,
                        10, 100, 10,
                        40, 3, 10,
                        0.01, 0.2, 1, 5,
                        TRUE, TRUE,
                        TRUE, 0.75, 0.75)
}
