% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create Configuration File From Arguments}
\usage{
create_config(outdir, type = "sc_3end", ...)
}
\arguments{
\item{outdir}{the destination directory for the configuratio nfile}

\item{type}{use an example config, available values:
\describe{
 \item{"sc_3end"}{ - config for 10x 3' end ONT reads}
 \item{"SIRV"}{ - config for the SIRV example reads}
}}

\item{...}{Configuration parameters.
\describe{
 \item{seed}{ - Integer. Seed for minimap2.}
 \item{threads}{ - Number of threads to use.}
 \item{do_barcode_demultiplex}{ - Boolean. Specifies whether to run the barcode demultiplexing step.}
 \item{do_genome_alignment}{ - Boolean. Specifies whether to run the genome alignment step. \code{TRUE} is recommended}
 \item{do_gene_quantification}{ - Boolean. Specifies whether to run gene quantification using the genome alignment results. \code{TRUE} is recommended}
 \item{do_isoform_identification}{ - Boolean. Specifies whether to run the isoform identification step. \code{TRUE} is recommended}
 \item{bambu_isoform_identification}{ - Boolean. Whether to use Bambu for isoform identification.}
 \item{multithread_isoform_identification}{ - Boolean. Whether to use FLAMES' new multithreaded Cpp implementation for isoform identification.}
 \item{do_read_realignment}{ - Boolean. Specifies whether to run the read realignment step. \code{TRUE} is recommended}
 \item{do_transcript_quantification}{ - Boolean. Specifies whether to run the transcript quantification step. \code{TRUE} is recommended}
 \item{barcode_parameters}{ - List. Parameters for barcode demultiplexing passed to \code{find_barcode} (except \code{fastq}, \code{barcodes_file}, \code{stats_out}, \code{reads_out}) and \code{threads}, which are set by the pipeline, see \code{?find_barcode} for more details.}
 \item{generate_raw_isoform}{ - Boolean. Whether to generate all isoforms for debugging purpose.}
 \item{max_dist}{ - Maximum distance allowed when merging splicing sites in isoform consensus clustering.}
 \item{max_ts_dist}{ - Maximum distance allowed when merging transcript start/end position in isoform consensus clustering.}
 \item{max_splice_match_dist}{ - Maximum distance allowed when merging splice site called from the data and the reference annotation.}
 \item{min_fl_exon_len}{ - Minimum length for the first exon outside the gene body in reference annotation. This is to correct the alignment artifact}
 \item{max_site_per_splice}{ - Maximum transcript start/end site combinations allowed per splice chain}
 \item{min_sup_cnt}{ - Minimum number of read support an isoform decrease this number will significantly increase the number of isoform detected.}
 \item{min_cnt_pct}{ - Minimum percentage of count for an isoform relative to total count for the same gene.}
 \item{min_sup_pct}{ - Minimum percentage of count for an splice chain that support a given transcript start/end site combination.}
 \item{strand_specific}{ - 0, 1 or -1. 1 indicates if reads are in the same strand as mRNA, -1 indicates reads are reverse complemented, 0 indicates reads are not strand specific.}
 \item{remove_incomp_reads}{ - The strenge of truncated isoform filtering. larger number means more stringent filtering.}
 \item{use_junctions}{ - whether to use known splice junctions to help correct the alignment results}
 \item{no_flank}{ - Boolean. for synthetic spike-in data. refer to Minimap2 document for detail}
 \item{use_annotation}{ - Boolean. whether to use reference to help annotate known isoforms}
 \item{min_tr_coverage}{ - Minimum percentage of isoform coverage for a read to be aligned to that isoform}
 \item{min_read_coverage}{ - Minimum percentage of read coverage for a read to be uniquely aligned to that isoform}
}}
}
\value{
file path to the config file created
}
\description{
Create Configuration File From Arguments
}
\details{
Create a list object containing the arguments supplied in a format usable for the FLAMES pipeline.
Also writes the object to a JSON file, which is located with the prefix 'config_' in the supplied \code{outdir}.
Default values from \code{extdata/config_sclr_nanopore_3end.json} will be used for unprovided parameters.
}
\examples{
# create the default configuration file
outdir <- tempdir()
config <- create_config(outdir)

}
