% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{find_isoform}
\alias{find_isoform}
\title{Title}
\usage{
find_isoform(
  anno,
  genome_bam,
  isoform_gff3,
  tss_tes_stat,
  genomefa,
  do_isoform_identification
)
}
\arguments{
\item{name}{desc}
}
\description{
DESC
}
