% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{find_isoform}
\alias{find_isoform}
\title{Isoform identification}
\usage{
find_isoform(annotation, genome_fa, genome_bam, outdir, config)
}
\arguments{
\item{annotation}{Path to annotation file. If configured to use bambu, the annotation
must be provided as GTF file.}

\item{genome_fa}{The file path to genome fasta file.}

\item{genome_bam}{File path to BAM alignment file. Multiple files could be provided.}

\item{outdir}{The path to directory to store all output files.}

\item{config}{Parsed FLAMES configurations.}
}
\value{
The updated annotation and the transcriptome assembly will be saved in the
output folder as \code{isoform_annotated.gff3} (GTF if bambu is selected) and
\code{transcript_assembly.fa} respectively.
}
\description{
Long-read isoform identification with FLAMES or bambu.
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
file_url <- "https://raw.githubusercontent.com/OliverVoogd/FLAMESData/master/data"
fastq1 <- bfc[[names(BiocFileCache::bfcadd(bfc, "Fastq1", paste(file_url, "fastq/sample1.fastq.gz", sep = "/")))]]
genome_fa <- bfc[[names(BiocFileCache::bfcadd(bfc, "genome.fa", paste(file_url, "SIRV_isoforms_multi-fasta_170612a.fasta", sep = "/")))]]
annotation <- bfc[[names(BiocFileCache::bfcadd(bfc, "annot.gtf", paste(file_url, "SIRV_isoforms_multi-fasta-annotation_C_170612a.gtf", sep = "/")))]]
outdir <- tempfile()
dir.create(outdir)
if (all(is.character(sys_which(c("minimap2", "k8"))))) {
    config <- jsonlite::fromJSON(system.file("extdata/SIRV_config_default.json", package = "FLAMES"))
    minimap2_align(
        config = config,
        fa_file = genome_fa,
        fq_in = fastq1,
        annot = annotation,
        outdir = outdir
    )
\dontrun{
    find_isoform(
        annotation = annotation, genome_fa = genome_fa,
        genome_bam = file.path(outdir, "align2genome.bam"),
        outdir = outdir, config = config
    )
}
}
}
