% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_measures}
\alias{fit_measures}
\title{LPML, DIC and WAIC measures}
\usage{
fit_measures(y, p)
}
\arguments{
\item{y}{Vector of response variable}

\item{p}{Matrix of probabilities for each individual y in each MCMC iteration}
}
\value{
measures A data.frame with LMPL, DIC and WAIC measures
}
\description{
LPML, DIC and WAIC measures
}
