% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_pipeline.R
\name{generic_long_pipeline}
\alias{generic_long_pipeline}
\title{Generic FLAMES pipeline}
\usage{
generic_long_pipeline(
  annotation,
  fastq,
  genome_bam,
  outdir,
  genome_fa,
  minimap2_dir,
  config
)
}
\arguments{
\item{annotation}{The file path to gene annotations file in gff3  format}

\item{fastq}{The file path to input fastq file}

\item{genome_bam}{optional BAM file which replaces fastq directory argument. This skips the genome alignment and
realignment steps}

\item{outdir}{The path to directory to store all output files.}

\item{genome_fa}{The file path to genome fasta file.}

\item{minimap2_dir}{Path to the directory containing minimap2, if it is not in PATH. Only required if either or both of
\code{do_genome_align} and \code{do_read_realign} are \code{TRUE}.}
}
\value{
This generic function returns a named list containing the output file names of the provided output files
in the given `outdir` directory. These files are loaded into R in either
a SummarizedExperiment or SingleCellExperiment object by the callers to this
function, `sc_long_pipeline()` and `bulk_long_pipeline()` respectively.
}
\description{
Generic implementation of the flames pipeline. Used for both bulk reads and
single cell reads.
}
