% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_pipeline.R
\name{generic_long_pipeline}
\alias{generic_long_pipeline}
\title{Generic FLAMES pipeline}
\usage{
generic_long_pipeline(
  annot,
  infq,
  in_bam,
  outdir,
  genome_fa,
  minimap2_dir,
  downsample_ratio,
  config_file,
  do_genome_align,
  do_isoform_id,
  do_read_realign,
  do_transcript_quanti,
  gen_raw_isoform,
  has_UMI,
  MAX_DIST,
  MAX_TS_DIST,
  MAX_SPLICE_MATCH_DIST,
  min_fl_exon_len,
  Max_site_per_splice,
  Min_sup_cnt,
  Min_cnt_pct,
  Min_sup_pct,
  strand_specific,
  remove_incomp_reads,
  use_junctions,
  no_flank,
  use_annotation,
  min_tr_coverage,
  min_read_coverage,
  bc_file = NULL
)
}
\arguments{
\item{annot}{gene annotations file in gff3  format}

\item{in_bam}{aligned bam file (sorted and indexed). If supplied, this overwrites
\code{fastq_dir} and skips the first alignment step}

\item{outdir}{directory to store all output files.}

\item{genome_fa}{genome fasta file.}

\item{minimap2_dir}{directory containing minimap2, k8 and paftools.js program.
k8 and paftools.js are used to convert gff3 to bed12.}

\item{downsample_ratio}{downsampling ratio if performing downsampling analysis.}

\item{config_file}{JSON configuration file. If specified, \code{config_file} overrides
all configuration parameters}

\item{do_genome_align}{Boolean. Specifies whether to run the genome alignment step. \code{TRUE} is recommended}

\item{do_isoform_id}{Boolean. Specifies whether to run the isoform identification step. \code{TRUE} is recommended}

\item{do_read_realign}{Boolean. Specifies whether to run the read realignment step. \code{TRUE} is recommended}

\item{do_transcript_quanti}{Boolean. Specifies whether to run the transcript quantification step. \code{TRUE} is recommended}

\item{has_UMI}{Boolean. Speficies if each gene as a UMI.}

\item{MAX_DIST}{Numeric}

\item{MAX_TS_DIST}{Numeric.}

\item{MAX_SPLICE_MATCH_DIST}{Numeric.}

\item{min_fl_exon_len}{Numeric.}

\item{Max_site_per_splice}{Numeric.}

\item{Min_sup_cnt}{Numeric.}

\item{Min_cnt_pct}{Numeric.}

\item{Min_sup_pct}{Numeric.}

\item{remove_incomp_reads}{Numeric.}

\item{use_junctions}{Boolean.}

\item{no_flank}{Boolean.}

\item{use_annotation}{Boolean.}

\item{min_tr_coverage}{Numeric.}

\item{min_read_coverage}{Numeric.}

\item{bc_file}{file containing the pseudo barcode annotations generated
from bulk_long_pipeline. If given, it is used for quantification.}
}
\description{
Generic implementation of the flames pipeline. Used for both bulk reads and
single cell reads.
}
