% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{get_gene_blocks}
\alias{get_gene_blocks}
\title{Get Gene Blocks}
\usage{
get_gene_blocks(gene_dict, chr_to_gene, gene_to_transcript)
}
\arguments{
\item{gene_dict}{NEEDED}

\item{chr_to_gene}{NEEDED}

\item{gene_to_transcript}{NEEDED}
}
\value{
NEEDED
}
\description{
NEEDED
}
\details{
NEEDED
}
