% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_to_fa.R
\name{get_transcript_seq}
\alias{get_transcript_seq}
\title{Get Transcript Sequences}
\usage{
get_transcript_seq(
  fa_file,
  fa_out_f,
  chr_to_gene,
  transcript_dict,
  gene_to_transcript,
  transcript_to_exon,
  ref_dict = NULL
)
}
\arguments{
\item{fa_out_f}{NEEDED}

\item{chr_to_gene}{NEEDED}

\item{transcript_dict}{NEEDED}

\item{gene_to_transcript}{NEEDED}

\item{transcript_to_exon}{NEEDED}

\item{ref_dict}{NEEDED}
}
\description{
NEEDED
}
