% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_to_fa.R
\name{get_transcript_seq}
\alias{get_transcript_seq}
\title{Get Transcript Sequences}
\usage{
get_transcript_seq(
  fa_file,
  fa_out_f,
  chr_to_gene,
  transcript_dict,
  gene_to_transcript,
  transcript_to_exon,
  ref_dict = NULL
)
}
\value{
returns NULL
}
\description{
Get Transcript Sequences
}
\examples{
genomefa <- system.file("extdata/SIRV_genomefa.fasta", package="FLAMES")
gff3_parse <- parse_gff_tree(system.file("extdata/isoform_annotated.gff3", package="FLAMES"))
get_transcript_seq(genomefa, tempfile(fileext=".fa"), gff3_parse$chr_to_gene, gff3_parse$transcript_dict, gff3_parse$gene_to_transcript, gff3_parse$transcript_to_exon)
}
