% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{gff3_to_bed12}
\alias{gff3_to_bed12}
\title{GFF3 to BED12}
\usage{
gff3_to_bed12(gff3_file, bed12_file)
}
\arguments{
\item{gff3_file}{The file path to the GFF3 file to convert}

\item{bed12_file}{The file path of the bed12 output file.}
}
\value{
file path to the created bed12_file.
}
\description{
Converts a gff3 file to a bed12
}
\examples{
annot <- system.file("extdata/SIRV_anno.gtf", package = "FLAMES")
out_bed <- tempfile(fileext = ".bed12")
\dontrun{
gff3_to_bed12(annot, out_bed)
}
}
