% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{group_bam2isoform}
\alias{group_bam2isoform}
\title{Group BAM to Isoform}
\usage{
group_bam2isoform(
  bam_in,
  out_gff3,
  out_stat,
  summary_csv,
  chr_to_blocks,
  gene_dict,
  transcript_to_junctions,
  transcript_dict,
  fa_f,
  config,
  downsample_ratio,
  raw_gff3 = NULL
)
}
\arguments{
\item{bam_in}{Input BAM file}

\item{out_gff3}{Output GFF3 file}

\item{out_stat}{NEEDED}

\item{summary_csv}{NEEDED}

\item{chr_to_blocks}{NEEDED}

\item{gene_dict}{NEEDED}

\item{transcript_to_junctions}{NEEDED}

\item{transcript_dict}{NEEDED}

\item{fa_f}{NEEDED}

\item{config}{Config file used to specify additional parameters}

\item{raw_gff3}{}
}
\value{
File paths of the output files \code{out_gff3} and \code{out_stat}
}
\description{
Short description of function
}
\details{
Further required details

\code{config} contains
\itemize{
\item MAX_DIST
\item MAX_TS_DIST
\item MAX_SPLICE_MATCH_DIST
\item Max_site_per_splice
\item Min_sup_cnt
\item min_fl_exon_len
\item Min_sup_pct
\item strand_specific
\item remove_incomp_reads
\item random_seed OPTIONAL;
}
}
