% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{group_bam2isoform}
\alias{group_bam2isoform}
\title{Group BAM to Isoform}
\usage{
group_bam2isoform(
  bam_in,
  out_gff3,
  out_stat,
  summary_csv,
  chr_to_blocks,
  gene_dict,
  transcript_to_junctions,
  transcript_dict,
  fa_f,
  config,
  downsample_ratio,
  raw_gff3 = NULL
)
}
\arguments{
\item{bam_in}{Input BAM file}

\item{out_gff3}{Output GFF3 file}

\item{out_stat}{NEEDED}

\item{summary_csv}{NEEDED}

\item{chr_to_blocks}{NEEDED}

\item{gene_dict}{NEEDED}

\item{transcript_to_junctions}{NEEDED}

\item{transcript_dict}{NEEDED}

\item{fa_f}{NEEDED}

\item{config}{Config file used to specify additional parameters}

\item{downsample_ratio}{NEEDED}

\item{raw_gff3}{NEEDER}
}
\value{
File paths of the output files \code{out_gff3} and \code{out_stat}
}
\description{
Short description of function
}
\details{
Further required details

\code{config} contains
\itemize{
    \item MAX_DIST
    \item MAX_TS_DIST
    \item MAX_SPLICE_MATCH_DIST
    \item Max_site_per_splice
    \item Min_sup_cnt
    \item min_fl_exon_len
    \item Min_sup_pct
    \item strand_specific
    \item remove_incomp_reads
    \item random_seed OPTIONAL;
}
}
\examples{
anno <- system.file("extdata/SIRV_anno.gtf", package="FlamesR")
genome_bam <- system.file("extdata/align2genome.bam", package="FlamesR")
genomfa <- system.file("extdata/SIRV_genomefa.fasta", package="FlamesR")
config <- parse_json_config(system.file("extdata/SIRV_config_default.json", package="FlamesR"))
gff3_parse <- parse_gff_tree(anno)
gene_dict <- get_gene_flat(gff3_parse$gene_to_transcript, gff3_parse$transcript_to_exon)
chr_to_blocks <- get_gene_blocks(gene_dict, gff3_parse$chr_to_gene, gff3_parse$gene_to_transcript)

juncs = list()
for (tr in names(gff3_parse$transcript_to_exon)) juncs[[tr]] = blocks_to_junctions(gff3_parse$transcript_to_exon[[tr]])
group_bam2isoform(genome_bam, out_gff3=tempfile(fileext=".gff3"), out_stat=tempfile(fileext=".bedgraph"),
                 "", chr_to_blocks, gene_dict, juncs, gff3_parse$transcript_dict, 
                 genomefa, config$isoform_parameters, 1, NULL)
}
