% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lambda_posterior}
\alias{lambda_posterior}
\title{Lambda for degree of freedom posterior distribution}
\usage{
lambda_posterior(p_df, p_lambda, inv_link_f, log = TRUE,
  method = "ARMS", a_lambda = NULL, b_lambda = NULL)
}
\arguments{
\item{p_df}{Degrees of freedom}

\item{p_lambda}{Lambda hyperparameter for p_df}

\item{inv_link_f}{Inverse link function}

\item{log}{For log-scale values}

\item{method}{"ARMS" or "metropolis"}
}
\value{
Likelihood value
}
\description{
Lambda for degree of freedom posterior distribution
}
