% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link}
\alias{link}
\title{Link function for several models}
\usage{
link(x, type = "logistic", df = 1)
}
\arguments{
\item{x}{Probability to evaluate the quantile}

\item{type}{"logit", "probit", "cauchit", "tobit" or "cloglog"}

\item{df}{Degrees of freedom (if type = "tobit")}
}
\value{
Quantile of x
}
\description{
Link function for several models
}
