% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{locate_minimap2_dir}
\alias{locate_minimap2_dir}
\title{locate parent folder of minimap2}
\usage{
locate_minimap2_dir(minimap2_dir = NULL)
}
\arguments{
\item{minimap2_dir}{(Optional) folder that includes minimap2}
}
\value{
Path to the parent folder of minimap2 if available, or
FALSE if minimap2 could not be found.
}
\description{
locate minimap2, or validate that minimap2 exists under minimap2_dir
}
\examples{
locate_minimap2_dir()
}
