% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_cell_barcode_cpp}
\alias{match_cell_barcode_cpp}
\title{Match Cell Barcodes}
\usage{
match_cell_barcode_cpp(
  fastq_dir,
  stats_file,
  out_fastq,
  ref_csv,
  MAX_DIST,
  UMI_LEN = 10L
)
}
\arguments{
\item{fastq_dir}{directory containing fastq files to match}

\item{stats_file}{NEEDED}

\item{out_fastq}{output filename for matched barcodes}

\item{ref_csv}{NEEDED}

\item{MAX_DIST}{int; maximum edit distance}

\item{UMI_LEN}{int; length of UMI sequences}
}
\value{
returns NULL
}
\description{
Match cell barcodes in the given fastq directory with the reference csv, \code{ref_csv}. Matches are returned
in the output file \code{out_fastq}
}
