% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_bin_metropolis.R
\name{mcmc_bin_metropolis}
\alias{mcmc_bin_metropolis}
\title{MCMC for binary regression - Metropolis version}
\usage{
mcmc_bin_metropolis(y, X, sample_size, burnin, inv_link_f, type, sample_c,
  sigma_beta, a_c, b_c, a_lambda, b_lambda, var_df, var_c, var_lambda, p_c,
  p_prop, p_beta, p_df, p_lambda)
}
\arguments{
\item{y}{Bernoulli observed values}

\item{X}{Covariate matrix}

\item{sample_size}{Sample size required for MCMC}

\item{burnin}{Burn in period (to adapt metropolis)}

\item{inv_link_f}{Inverse link function}

\item{type}{"logit", "probit", "cauchit", "robit" or "cloglog"}

\item{sample_c}{Should c be sampled?}

\item{sigma_beta}{Variance of beta prior}

\item{a_c}{Shape1 for c prior}

\item{b_c}{Shape2 for c prior}

\item{a_lambda}{Inferior limit for lambda}

\item{b_lambda}{Superior limit for lambda}

\item{var_df}{Variance to sample log(df)}

\item{var_c}{Variance to sample c}

\item{var_lambda}{Variance to sample lambda}

\item{p_c}{To restore c}

\item{p_prop}{To restore p}

\item{p_beta}{To restore beta}

\item{p_df}{To restore df}

\item{p_lambda}{To restore lambda}
}
\value{
Chains of all parameters
}
\description{
Perform MCMC with Metropolis Hastings algorithm
}
