% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{minimap2_align}
\alias{minimap2_align}
\title{Minimap2 Align to Genome}
\usage{
minimap2_align(
  minimap2_prog_path = NULL,
  fa_file,
  fq_in,
  bam_out,
  no_flank = FALSE,
  bed12_junc = NULL
)
}
\arguments{
\item{minimap2_prog_path}{Absolute path to the directory containing k8, paftools,js and minimap2}

\item{fa_file}{d}

\item{fq_in}{d}

\item{bam_out}{Output BAM file}

\item{no_flank}{d}

\item{bed12_junc}{d}
}
\description{
DESC
}
