% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{minimap2_align}
\alias{minimap2_align}
\title{Minimap2 Align to Genome}
\usage{
minimap2_align(
  minimap2_prog_path = NULL,
  fa_file,
  fq_in,
  sam_out,
  no_flank = FALSE,
  bed12_junc = NULL
)
}
\arguments{
\item{minimap2_prog_path}{Absolute path to the directory containing minimap2}

\item{fa_file}{Path to the fasta file used as a reference database for alignment}

\item{fq_in}{File path to the fastq file used as a query sequence file}

\item{sam_out}{File path to the output SAM file}

\item{no_flank}{Boolean; used if studying SIRV, to tell minimap2 to ignore additional bases}

\item{bed12_junc}{file path to the gene annotations in BED12 format. If specified, minmap2 prefers splicing in annotations.}
}
\value{
file path to the given output BAM file, \code{bam_out}
}
\description{
Uses minimap2 to align sequences agains a reference databse.
Uses options "-ax splice -t 12 -k14 --secondary=no \code{fa_file} \code{fq_in}"
}
