% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gene_anno.R
\name{parse_gff_tree}
\alias{parse_gff_tree}
\title{Parse GFF3 file formats}
\usage{
parse_gff_tree(gff_file)
}
\arguments{
\item{gff_file}{File path in gff, gtf.gz or gtf format to parse.}
}
\description{
Parses GFF3 file formats into a usable list of four elements. Accepts \code{.gtf}, \code{.gff} formats as 
well as \code{.gz} compressed versions.
}
\examples{
gff3_parse <- parse_gff_tree(system.file("extdata/SIRV_anno.gtf", package="FLAMES"))
}
