% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gene_anno.R
\name{parse_gff_tree}
\alias{parse_gff_tree}
\title{Parse Gff3 file}
\usage{
parse_gff_tree(gff_file)
}
\arguments{
\item{gff_file}{the file path to the gff3 file to parse}
}
\value{
a named list with the elements
"chr_to_gene", "transcript_dict", "gene_to_transcript", "transcript_to_exon", containing
the data parsed from the gff3 file.
}
\description{
Parse a Gff3 file into 3 components: chromasome to gene name, a transcript dictionary, a gene to transcript dictionary
and a transcript to exon dictionary.
These components are returned in a named list.
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
file_url <-
    "https://raw.githubusercontent.com/OliverVoogd/FLAMESData/master/data"
gff <- bfc[[names(BiocFileCache::bfcadd(bfc, "GFF", paste(file_url, "SIRV_isoforms_multi-fasta-annotation_C_170612a.gtf", sep = "/")))]]
\donttest{
parsed_gff <- parse_gff_tree(gff)
}
}
