% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{parse_json_config}
\alias{parse_json_config}
\title{Parse Json Configuration File}
\usage{
parse_json_config(json_file)
}
\arguments{
\item{json_file}{The file name to convert into an R list.}
}
\value{
A named R list of the parameters in \code{json_file}. Subsections are: \code{pipeline_parameters},
     \code{global_parameters}, \code{isoform_parameters}, \code{alignment_parameters}, \code{realign_parameters} and
     \code{transcript_counting}.
}
\description{
Convert a json configuration file into a named R list, grouped into sub lists according to their
     usage in the Flames pipeline.
}
