% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{parse_json_config}
\alias{parse_json_config}
\title{Parse Json Configuration File}
\usage{
parse_json_config(json_file)
}
\arguments{
\item{json_file}{the file path to the JSON file to convert into an R list. 
This can be the default FLAMES configuration file found using \code{get_default_config_file()}}
}
\value{
A named R list of the parameters in \code{json_file}. Subsections are: \code{pipeline_parameters},
     \code{global_parameters}, \code{isoform_parameters}, \code{alignment_parameters}, \code{realign_parameters} and
     \code{transcript_counting}.
}
\description{
Convert a json configuration file into a named R list, grouped into sub lists according to their
     usage in the Flames pipeline.
}
\examples{
config <- get_default_config_file()
\dontrun{
parse_json_config(config)
}
}
