% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tr.R
\name{parse_realigned_bam}
\alias{parse_realigned_bam}
\title{Title}
\usage{
parse_realigned_bam(
  bam_in,
  fa_idx_f,
  min_sup_reads,
  min_tr_coverage,
  min_read_coverage,
  ...
)
}
\arguments{
\item{name}{desc}
}
\description{
DESC
}
