% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tr.R
\name{parse_realigned_bam}
\alias{parse_realigned_bam}
\title{Parse Realigned Bam}
\usage{
parse_realigned_bam(
  bam_in,
  fa_idx_f,
  min_sup_reads,
  min_tr_coverage,
  min_read_coverage,
  ...
)
}
\arguments{
\item{bam_in}{NEEDED}

\item{fa_idx_f}{NEEDED}

\item{min_sup_reads}{NEEDED}

\item{min_tr_coverage}{NEEDED}

\item{min_read_coverage}{NEEDED}

\item{...}{Additional parameters, which can include the file \code{bc_file}
to serve as a dictionary of barcode as reference}
}
\value{
named list containing the elements \code{bc_tr_count_dict}, \code{bc_tr_badcov_count_dict} and \code{tr_kept}
}
\description{
NEEDED
}
\examples{

realign_bam <- system.file("extdata/align2genome.bam", package="FlamesR")
transcript_fa <- system.file("extdata/transcript_assembly.fa.fai", package="FlamesR")
\dontrun{
parse_realign <- parse_realigned_bam(realign_bam, transcript_fa, 10, 0.75, 0.75)
}
}
