% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_decay.R
\name{plot_coverage}
\alias{plot_coverage}
\title{plot read coverages}
\usage{
plot_coverage(
  bam,
  isoform = NULL,
  quantiles = c(0, 0.2375, 0.475, 0.7125, 0.95, 1),
  length_bins = c(0, 1, 2, 5, 10, Inf),
  weight_fn = "read_counts",
  annotation,
  coding_only = FALSE,
  remove_utr = FALSE
)
}
\arguments{
\item{bam, }{path to the BAM file (aligning reads to the transcriptome), or
the (GenomicAlignments::readGAlignments) parsed GAlignments object}

\item{isoform}{string vector, provide isoform names to plot the coverage for the
corresponding isoforms, or provide NULL to plot average coverages for each
length bin}

\item{quantiles}{numeric vector to specify the quantiles to bin the transcripts lengths
by if length_bins is missing. The length bins will be determined such that the read
counts are distributed acording to the quantiles.}

\item{length_bins, }{numeric vector to specify the sizes to bin the isoforms by}

\item{weight_fn}{"read_counts" or "sigmoid", determins how the transcripts
should be weighted within length bins.}

\item{annotation}{path to the GTF annotation file, or the parsed txdb object}

\item{coding_only}{logical, if TRUE, only plot the coverage for coding transcripts}

\item{remove_utr}{logical, if TRUE, trim the UTRs from the transcripts}
}
\value{
a ggplot2 object of the coverage plot(s)
}
\description{
Plot the average read coverages for each length bin or a
perticular isoform
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
file_url <- 'https://raw.githubusercontent.com/OliverVoogd/FLAMESData/master/data'
fastq1 <- bfc[[names(BiocFileCache::bfcadd(bfc, 'Fastq1', paste(file_url, 'fastq/sample1.fastq.gz', sep = '/')))]]
genome_fa <- bfc[[names(BiocFileCache::bfcadd(bfc, 'genome.fa', paste(file_url, 'SIRV_isoforms_multi-fasta_170612a.fasta', sep = '/')))]]
annotation <- bfc[[names(BiocFileCache::bfcadd(bfc, 'annot.gtf', paste(file_url, 'SIRV_isoforms_multi-fasta-annotation_C_170612a.gtf', sep = '/')))]]
outdir <- tempfile()
dir.create(outdir)
fasta <- annotation_to_fasta(annotation, genome_fa, outdir)
minimap2_realign(
  config = jsonlite::fromJSON(
    system.file("extdata", "config_sclr_nanopore_3end.json", package = "FLAMES")),
  fq_in = fastq1,
  outdir = outdir
)
plot_coverage(bam = file.path(outdir, 'realign2transcript.bam'))
}
