% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread_functions.R
\name{remove_similar_tr}
\alias{remove_similar_tr}
\title{Remove Similar Transcripts}
\usage{
remove_similar_tr(
  transcript_dict,
  gene_to_transcript,
  transcript_to_exon,
  thr = 10
)
}
\arguments{
\item{name}{desc}
}
\description{
DESC
}
