% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_longread.R
\name{remove_similar_tr}
\alias{remove_similar_tr}
\title{Remove Similar Transcripts}
\usage{
remove_similar_tr(gene_to_transcript, transcript_to_exon, thr = 10)
}
\arguments{
\item{gene_to_transcript}{NEEDED}

\item{transcript_to_exon}{NEEDED}

\item{thr}{The threshold for exon similarity}
}
\value{
return NULL
}
\description{
Remove any duplicate exons, or similar exons with a similarity greater than
\code{thr}, from each element in \code{gene_to_transcript}. Modifies
\code{gene_to_transcript} in order to remove duplicates.
}
\examples{
gff3_parse <- parse_gff_tree(system.file("extdata/SIRV_anno.gtf", package="FlamesR"))
remove_similar_tr(gff3_parse$gene_to_transcript, gff3_parse$transcript_to_exon, thr=5)
}
