% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{samtools_as_bam}
\alias{samtools_as_bam}
\title{Samtools as BAM}
\usage{
samtools_as_bam(sam_in, bam_out)
}
\arguments{
\item{sam_in}{the input SAM file}

\item{bam_out}{the output BAM file}
}
\value{
the same value as \code{bam_out}
}
\description{
Produces a compressed binary BAM file from a text based SAM file, using Rsamtools.
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask=FALSE)
file_url <- 
   "https://raw.githubusercontent.com/OliverVoogd/FLAMESData/master/data"
tmp_sam <- paste0(temp_path, "/tmp_sam.sam")
file.rename(bfc[[names(BiocFileCache::bfcadd(bfc, "Temp SAM", paste(file_url, "align2genome.bam", sep="/")))]], genome_bam)

samtools_as_bam(tmp_sam, tempfile(fileext="bam"))
}
