% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{samtools_sort_index}
\alias{samtools_sort_index}
\title{Samtools Sort Index}
\usage{
samtools_sort_index(bam_in, bam_out)
}
\arguments{
\item{bam_in}{Input BAM File; MORE}

\item{name}{Output BAM file; MORE}
}
\description{
Sorts and then indexs alignments for fast random access using samtools.
}
