% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{samtools_sort_index}
\alias{samtools_sort_index}
\title{Samtools Sort Index}
\usage{
samtools_sort_index(bam_in, bam_out)
}
\arguments{
\item{bam_in}{Input BAM File}

\item{bam_out}{Output BAM file}
}
\value{
the path to the output file, given as \code{bam_out}
}
\description{
Sorts and then indexes alignments for fast random access using samtools. Requires samtools to 
be in PATH
}
