% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{samtools_sort_index}
\alias{samtools_sort_index}
\title{Samtools Sort and Index}
\usage{
samtools_sort_index(bam_in, bam_out)
}
\arguments{
\item{bam_in}{the bam file to sort and index}

\item{bam_out}{the output indexed bam file.}
}
\value{
the same value as \code{bam_out}
}
\description{
Sort and index the given BAM file, using Rsamtools.
}
\examples{
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
file_url <-
    "https://raw.githubusercontent.com/OliverVoogd/FLAMESData/master/data"
genome_bam <- paste0(temp_path, "/align2genome.bam")
file.rename(bfc[[names(BiocFileCache::bfcadd(bfc, "Genome BAM", paste(file_url, "align2genome.bam", sep = "/")))]], genome_bam)

samtools_sort_index(genome_bam, tempfile(fileext = ".bam"))
}
