% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_annotate_plots.R
\name{sc_annotate_plots}
\alias{sc_annotate_plots}
\title{FLAMES Annotated Plottings}
\usage{
sc_annotate_plots(
  gene,
  multiAssay,
  cluster_annotation,
  n_isoforms = 4,
  n_pcs = 40,
  return_multiAssay = TRUE,
  heatmap_annotation_colors = "BrBG",
  isoform_legend_width = 7,
  heatmap_color_quantile = 0.95,
  col_low = "#313695",
  col_mid = "#FFFFBF",
  col_high = "#A50026"
)
}
\arguments{
\item{gene}{The gene symbol of interest.}

\item{multiAssay}{The \code{MultiAssayExperiment} object from \code{combine_sce()}.}

\item{cluster_annotation}{Path to the cluster annotation CSV (required for heatmap, if \code{cluster_annotation.csv} is not in \code{path} and \code{multiAssay$cell_type} does not exist)}

\item{n_isoforms}{The number of expressed isoforms to keep.}

\item{n_pcs}{The number of principal components to generate.}

\item{return_multiAssay}{Whether to return the processed \code{MultiAssayExperiment} object.}

\item{heatmap_annotation_colors}{Name of color palette to use for cell group annotation in heatmaps, see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}} \cr
available diverging palettes are:\cr
\code{BrBG PiYG PRGn PuOr RdBu RdGy RdYlBu RdYlGn Spectral}\cr
when there are more than 11 groups, this argument will be ignored and random palettes will be generated.}

\item{isoform_legend_width}{The width of isoform legends in heatmaps, in \code{cm}.}

\item{heatmap_color_quantile}{Float; Expression levels higher than this quantile will all be shown with \code{col_high}.
Expression levels lower than 1 - \code{heatmap_color_quantile} will all be shown with \code{col_low};}

\item{col_low}{Color for cells with low expression levels in UMAPs.}

\item{col_mid}{Color for cells with intermediate expression levels in UMAPs.}

\item{col_high}{Color for cells with high expression levels in UMAPs.}
}
\value{
a list containing the combined UMAP, the isoform exon alignments and the UMAP with isoform expression levels.
}
\description{
Plot isoform exons alignments for a given gene, along with UMAP showing expression levels.
}
\details{
This function takes a combined \code{MultiAssayExperiment} object containing "gene_counts" and "transcript_counts" experiments
to generate a combined UMAP, the expression levels of isoforms (using long read data)
are then overlayed on top of the UMAP. SNN inference based on gene counts were performed to impute isoform expression
for cells in the larger library. The \code{MultiAssayExperiment} object should have a boolean columen named "Lib_small" in
its columen data file to idicate which subsample the cells are in. The \code{MultiAssayExperiment} object can be created using
the \code{combine_sce} function.
}
\seealso{
\code{\link[=combine_sce]{combine_sce()}} for combining gene count \code{SingleCellExperiment} object with transcript counts.
}
