% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_long_pipeline.R
\name{sc_long_pipeline}
\alias{sc_long_pipeline}
\title{Title}
\usage{
sc_long_pipeline(
  annot,
  infq,
  inbam = NULL,
  outdir,
  genomefa,
  minimap2_dir = NULL,
  config_file = NULL,
  downsample_ratio = 1
)
}
\arguments{
\item{name}{desc}
}
\description{
DESC
}
