% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_mutations.R
\name{sc_mutations}
\alias{sc_mutations}
\title{FLAMES variant calling}
\usage{
sc_mutations(
  fa_f,
  bam_short,
  out_dir,
  barcode_tsv,
  gff_f = NULL,
  known_positions = NULL,
  min_cov = 100,
  report_pct = c(0.15, 0.85),
  test_mode = FALSE
)
}
\arguments{
\item{fa_f}{Reference genome FASTA file}

\item{bam_short}{Optional short read alignment BAM file}

\item{out_dir}{Output folder of sc_long_pipeline. Output files from this function will also be saved here.}

\item{gff_f}{Reference genome annotation file, in \code{GFF} or \code{GTF} format. If left with \code{NONE} then the \code{isoform_annotated.gff3} from \code{sc_longread_pipeline} will be used.}

\item{known_positions}{A list of known positions, with by chromosome name followed by the position, e.g. ('chr1', 123, 'chr1', 124, 'chrX', 567)}

\item{min_cov}{The coverage threshold for filtering candidate SNVs}

\item{report_pct}{The allele frequency range for filtering candidate SNVs}

\item{barcode}{TSV file for cell barcodes}
}
\value{
a \code{data.frame} containing the following columns:
\itemize{
 \item{chr}{ - the chromosome where the mutation is located}
 \item{position}
 \item{REF}{ - the reference allele}
 \item{ALT}{ - the alternative allele}
 \item{REF_frequency}{ - reference allele frequency}
 \item{REF_frequency_in_short_reads}{ - reference allele frequency in short reads (-1 when short reads not provided)}
 \item{hypergeom_test_p_value}
 \item{sequence_entrophy}
 \item{INDEL_frequency}
 \item{adj_p}{ - the adjusted p-value (by Benjamini–Hochberg correction)}
}
The table is sorted by decreasing adjusted P value.

files saved to out_dir/mutation:
\itemize{
     \item{ref_cnt.csv.gz}
     \item{alt_cnt.csv.gz}
     \item{allele_stat.csv.gz}
     \item{freq_summary.csv}
}
}
\description{
Candidate SNVs identified with filtering by coverage threshold,
and allele frequency range.
}
