% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_mutations.R
\name{sc_mutations}
\alias{sc_mutations}
\title{FLAMES variant calling}
\usage{
sc_mutations(
  sce,
  barcode_tsv,
  bam_short,
  out_dir,
  genome_fa,
  annot,
  known_positions = NULL,
  min_cov = 100,
  report_pct = c(0.15, 0.85),
  test_mode = FALSE
)
}
\arguments{
\item{sce}{The \code{SingleCellExperiment} object from \code{sc_long_pipeline}}

\item{bam_short}{(Optional) short read alignment BAM file}

\item{out_dir}{(Optional) Output folder of sc_long_pipeline. Output files from this function will also be saved here.
Use this parameter if you do not have the \code{SingleCellExperiment} object.}

\item{genome_fa}{(Optional) Reference genome FASTA file. Use this parameter is if you do not wish \code{sc_mutation} to use the
reference genome FASTA file from the \code{sce}'s metadata.}

\item{annot}{(Optional) The file path to gene annotation file in gff3 format. If provided as \code{FALSE} then the \code{isoform_annotated.gff3} from \code{sc_longread_pipeline} will be used, if not provided then the path in the \code{SingleCellExperiment} object will be used.}

\item{known_positions}{(Optional) A list of known positions, with by chromosome name followed by the position, e.g. ('chr1', 123, 'chr1', 124, 'chrX', 567)}

\item{min_cov}{The coverage threshod for filtering candidate SNVs}

\item{report_pct}{The allele frequency range for filtering candidate SNVs}

\item{barcode}{TSV file for cell barcodes}
}
\value{
a \code{data.frame} containing the following columns:
\itemize{
 \item{chr}{ - the chromosome where the mutation is located}
 \item{position}
 \item{REF}{ - the reference allele}
 \item{ALT}{ - the alternative allele}
 \item{REF_frequency}{ - reference allele frequency}
 \item{REF_frequency_in_short_reads}{ - reference allele frequency in short reads (-1 when short reads not provided)}
 \item{hypergeom_test_p_value}
 \item{sequence_entrophy}
 \item{INDEL_frequency}
 \item{adj_p}{ - the adjusted p-value (by Benjamini–Hochberg correction)}
}
The table is sorted by decreasing adjusted P value.

files saved to out_dir/mutation:
\itemize{
     \item{ref_cnt.csv.gz}
     \item{alt_cnt.csv.gz}
     \item{allele_stat.csv.gz}
     \item{freq_summary.csv}
}
}
\description{
Candidate SNVs identified with filtering by coverage threshold,
and allele frequency range.
}
\details{
Takes the \code{SingleCellExperiment} object from \code{sc_long_pipeline} and the cell barcodes as \code{barcode}.
Alternatively, input can also be provided as \code{out_dir}, \code{genome_fa}, \code{annot}, \code{barcode}.
}
