% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_pipelines.R
\name{windows_pipeline_isoforms}
\alias{windows_pipeline_isoforms}
\title{Windows Pipeline - Find Isoforms}
\usage{
windows_pipeline_isoforms(pipeline_variables)
}
\arguments{
\item{pipeline_variables}{the list returned from \code{windows_pipeline_isoforms}.}
}
\value{
the updated \code{pipeline_variables} list, with information required for the final pipeline step.
}
\description{
This is the second step in the 3 step Windows FLAMES pipeline. 
Following this step, read realignment should be undertaken, using the file paths
given in the return pipeline_variables$return_files.
After this has been completed, the final pipeline step, \code{windows_pipeline_quantification} should be run,
giving the returned list from this function as input.
}
