% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_pipelines.R
\name{windows_pipeline_quantification}
\alias{windows_pipeline_quantification}
\title{Windows Pipeline - Quantification}
\usage{
windows_pipeline_quantification(pipeline_vars)
}
\arguments{
\item{pipeline_variables}{the list returned from \code{windows_pipeline_isoforms}, containing the information
required to perform the final step, quantification.}
}
\value{
\code{windows_pipeline_quantification} returns a SummarizedExperiment object, or a SingleCellExperiment in the case
of this function being used for the FLAMES single cell pipeline, containing a count
matrix as an assay, gene annotations under metadata, as well as a list of the other
output files generated by the pipeline. The pipeline also outputs a number of output
files into the given \code{outdir} directory. These output files generated by the pipeline are:
\itemize{
 \item{transcript_count.csv.gz}{ - a transcript count matrix (also contained in the SummarizedExperiment)}
 \item{isoform_annotated.filtered.gff3}{ - isoforms in gff3 format (also contained in the SummarizedExperiment)}
 \item{transcript_assembly.fa}{ - transcript sequence from the isoforms}
 \item{align2genome.bam}{ - sorted BAM file with reads aligned to genome}
 \item{realign2transcript.bam}{ - sorted realigned BAM file using the transcript_assembly.fa as reference}
 \item{tss_tes.bedgraph}{ - TSS TES enrichment for all reads (for QC)}
}
}
\description{
This is the final step in the 3 step Windows FLAMES pipeline. This should be run
after read realignment is performed, following \code{windows_pipeline_isoforms}.
}
