% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tr.R
\name{wrt_tr_to_csv}
\alias{wrt_tr_to_csv}
\title{Write Transcript to CSV file}
\usage{
wrt_tr_to_csv(
  bc_tr_count_dict,
  transcript_dict,
  csv_f,
  transcript_dict_ref = NULL,
  has_UMI = TRUE
)
}
\arguments{
\item{bc_tr_count_dict}{NEEDED}

\item{transcript_dict}{NEEDED}

\item{csv_f}{NEEDED}

\item{transcript_dict_ref}{NEEDED}

\item{has_UMI}{NEEDED}
}
\value{
NEEDED
}
\description{
NEEDED
}
\examples{
isoform_gff3 <- parse_gff_tree(system.file("extdata/isoform_annotated.gff3", package="FLAMES"))
gff3_parse <- parse_gff_tree(system.file("extdata/SIRV_anno.gtf", package="FLAMES"))
realign_bam <- system.file("extdata/align2genome.bam", package="FLAMES")
\dontrun{
parse_realign <- parse_realigned_bam(realign_bam, system.file("extdata/transcript_assembly.fa.fai", package="FLAMES"), 10, 0.75, 0.75)
tr_cnt <- wrt_tr_to_csv(parse_realign$bc_tr_count_dict, isoform_gff3$transcript_dict, tempfile(fileext=".csv.gz"), 
                        gff3_parse$transcript_dict, FALSE)
                        }
}
