% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tr.R
\name{wrt_tr_to_csv}
\alias{wrt_tr_to_csv}
\title{Write Transcript to CSV file}
\usage{
wrt_tr_to_csv(
  bc_tr_count_dict,
  transcript_dict,
  csv_f,
  transcript_dict_ref = NULL,
  has_UMI = TRUE
)
}
\arguments{
\item{bc_tr_count_dict}{NEEDED}

\item{transcript_dict}{NEEDED}

\item{csv_f}{NEEDED}

\item{transcript_dict_ref}{NEEDED}

\item{has_UMI}{NEEDED}
}
\value{
NEEDED
}
\description{
NEEDED
}
